% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/adjust.R
\docType{methods}
\name{adjust}
\alias{adjust}
\alias{adjust-method}
\alias{adjust,GammaSpectra-method}
\alias{adjust,GammaSpectrum-method}
\title{Energy shift}
\usage{
adjust(object, ...)

\S4method{adjust}{GammaSpectra}(object, peaks = c(238, 1461, 2614.5),
  ...)

\S4method{adjust}{GammaSpectrum}(object, peaks = c(238, 1461, 2614.5),
  ...)
}
\arguments{
\item{object}{An object of class \linkS4class{GammaSpectra}.}

\item{...}{Currently not used.}

\item{peaks}{A \code{\link{numeric}} vector.}
}
\value{
An object of class \linkS4class{GammaSpectra}.
}
\description{
Adjusts spectrum.
}
\author{
N. Frerebeau
}
