% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AIX1}
\alias{AIX1}
\title{CEREGE Calibration Curve #1}
\format{An object of class \code{CalibrationCurve} of length 1.}
\usage{
AIX1
}
\description{
CEREGE Calibration Curve #1
}
\details{
\tabular{ll}{
 \strong{Laboratory:} \tab CEREGE \cr
 \strong{Instrument:} \tab Canberra Inspector 1000 \cr
 \strong{Detector:} \tab NaI \cr
 \strong{Authors:} \tab CEREGE Luminescence Team
}
}
\section{Changelog}{

 \describe{
  \item{v0.1.1}{Curve creation.
  Model summary:
  slope: 0.018 +/- 0;
  intercept: -34.361 +/- 49.326;
  residual standard error: 34.58;
  multiple R-squared: 0.99775;
  adjusted R-squared: 0.997}
 }
}

\examples{
# Load the curve
utils::data("AIX1", package = "gamma")
AIX1

# See how to replicate the curve
utils::vignette("AIX1#1", package = "gamma")
}
\keyword{datasets}
