% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AIX100_curve}
\alias{AIX100_curve}
\title{CEREGE Calibration Curve #1}
\format{An object of class \code{CalibrationCurve} of length 1.}
\usage{
data("AIX100_curve")
}
\description{
CEREGE Calibration Curve #1
}
\details{
\tabular{ll}{
 \strong{Laboratory} \tab CEREGE \cr
 \strong{Instrument} \tab Canberra Inspector 1000 \cr
 \strong{Detector} \tab NaI \cr
 \strong{Authors} \tab CEREGE Luminescence Team
}
}
\examples{
\donttest{
## Load the curve
utils::data("AIX100_curve", package = "gamma")
AIX100_curve

## See how to replicate the curve
utils::vignette("AIX1#1", package = "gamma")
}
}
\seealso{
Other datasets: \code{\link{.decay_La}},
  \code{\link{.decay}}, \code{\link{BDX100_curve}},
  \code{\link{BDX200_curve}}, \code{\link{BDX300_curve}},
  \code{\link{clermont}}
}
\concept{datasets}
\keyword{datasets}
