% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AIX_NaI_1}
\alias{AIX_NaI_1}
\title{CEREGE Calibration Curve (NaI)}
\format{
An object of class \linkS4class{CalibrationCurve}.
\tabular{ll}{
\strong{Laboratory} \tab CEREGE \cr
\strong{Instrument} \tab Canberra Inspector 1000 \cr
\strong{Detector} \tab NaI \cr
\strong{Authors} \tab CEREGE Luminescence Team
}
}
\usage{
data(AIX_NaI_1)
}
\description{
CEREGE Calibration Curve (NaI)
}
\examples{
## Load the curve
data(AIX_NaI_1, package = "gamma")
plot(AIX_NaI_1)
}
\seealso{
Other datasets: 
\code{\link{BDX_LaBr_1}},
\code{\link{HD_CZT_1}},
\code{\link{clermont}},
\code{\link{clermont_2024}},
\code{\link{heidelberg_2024}}
}
\concept{datasets}
\keyword{datasets}
