% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AIX_NaI_curve}
\alias{AIX_NaI_curve}
\title{CEREGE Calibration Curve (NaI)}
\format{
An object of class \linkS4class{CalibrationCurve}.
\tabular{ll}{
 \strong{Laboratory} \tab CEREGE \cr
 \strong{Instrument} \tab Canberra Inspector 1000 \cr
 \strong{Detector} \tab NaI \cr
 \strong{Authors} \tab CEREGE Luminescence Team
}
}
\usage{
data(AIX_NaI_curve)
}
\description{
CEREGE Calibration Curve (NaI)
}
\examples{
\donttest{
## Load the curve
utils::data(AIX_NaI_curve, package = "gamma")
AIX_NaI_curve

## View raw data
utils::data(AIX_NaI_data, package = "gamma")
AIX_NaI_data

## See how to replicate the curve
utils::vignette("AIX1#1", package = "gamma")
}
}
\seealso{
\link{AIX_NaI_data}

Other datasets: 
\code{\link{AIX_NaI_data}},
\code{\link{BDX_LaBr_1_curve}},
\code{\link{BDX_LaBr_1_data}},
\code{\link{clermont}}
}
\concept{datasets}
\keyword{datasets}
