% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AIX_NaI_data}
\alias{AIX_NaI_data}
\title{CEREGE Calibration Curve Data (NaI)}
\format{
An object of class \code{data.frame} with 5 rows and 10 columns.
}
\usage{
data(AIX_NaI_data)
}
\description{
CEREGE Calibration Curve Data (NaI)
}
\seealso{
\link{AIX_NaI_curve}

Other datasets: 
\code{\link{AIX_NaI_curve}},
\code{\link{BDX_LaBr_1_curve}},
\code{\link{BDX_LaBr_1_data}},
\code{\link{clermont}}
}
\concept{datasets}
\keyword{datasets}
