% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BaseLine-class}
\alias{BaseLine-class}
\alias{.BaseLine}
\title{An S4 Class to Represent a Spectrum Baseline}
\description{
An S4 Class to Represent a Spectrum Baseline
}
\note{
This class extends the \linkS4class{GammaSpectrum} class.
}
\examples{
## Import a CNF file
spc_file <- system.file("extdata/LaBr.CNF", package = "gamma")
spc <- read(spc_file)

## Remove the first 35 chanels
spc <- slice_signal(spc, -c(1:35))

## Estimate baseline
baseline <- estimate_baseline(spc)

plot(spc, baseline)

## Remove baseline
spc_clean1 <- remove_baseline(spc)
spc_clean2 <- spc - baseline

plot(spc_clean1)
}
\seealso{
\linkS4class{GammaSpectrum}.

Other class: 
\code{\link{DoseRateModel-class}},
\code{\link{GammaSpectra-class}},
\code{\link{GammaSpectrum-class}},
\code{\link{PeakPosition-class}}
}
\author{
N. Frerebeau
}
\concept{class}
