% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/baseline.R
\docType{methods}
\name{baseline}
\alias{baseline}
\alias{estimateBaseline}
\alias{estimateBaseline-method}
\alias{removeBaseline}
\alias{removeBaseline-method}
\alias{estimateBaseline,GammaSpectra-method}
\alias{estimateBaseline,GammaSpectrum-method}
\alias{removeBaseline,GammaSpectra-method}
\alias{removeBaseline,GammaSpectrum-method}
\title{Baseline estimation and removal}
\usage{
estimateBaseline(object, ...)

removeBaseline(object, ...)

\S4method{estimateBaseline}{GammaSpectra}(object, method = c("SNIP"),
  LLS = FALSE, ...)

\S4method{estimateBaseline}{GammaSpectrum}(object, method = c("SNIP"),
  LLS = FALSE, ...)

\S4method{removeBaseline}{GammaSpectra}(object, ...)

\S4method{removeBaseline}{GammaSpectrum}(object, ...)
}
\arguments{
\item{object}{A \code{\link{numeric}} vector.}

\item{...}{Extra parameters passed to other methods.}

\item{method}{TODO.}

\item{LLS}{A \code{\link{logical}} scalar: should the LLS operator be applied
on \code{x} before employing SNIP algorithm?}
}
\value{
\code{estimateBaseline} returns a four columns data frame or a list
 of data frame.

 \code{removeBaseline} returns a \linkS4class{GammaSpectrum} or a
 \linkS4class{GammaSpectra} object (same as \code{object}).
}
\description{
Baseline estimation and removal
}
\examples{
# Import a CNF file
spc_dir <- system.file("extdata/test1.cnf", package = "gamma")
spc_dir <- system.file("extdata/cerege", package = "gamma")
spc_test <- read(spc_dir)[[5]]

# Estimate baseline
baseline <- estimateBaseline(spc_test)

# Plot spectrum + baseline
plot(spc_test) +
  ggplot2::geom_line(colour = "red", data = baseline,
                     mapping = ggplot2::aes(x = energy, y = baseline))

# Remove baseline
spc_clean <- removeBaseline(spc_test)

# Detect peaks
peak_index <- findPeaks(spc_clean, span = NULL)

plot(spc_clean, select = 1:3, facet = TRUE) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts") +
  ggplot2::geom_vline(xintercept = peak_index$energy, colour = "red") +
  ggplot2::theme_bw()
}
\references{
Morhác, M., Kliman, J., Matoucek, V., Veselský, M. and Turzo, I. (1997).
 Background elimination methods for multidimensional gamma-ray spectra.
 \emph{NIM}, A401 (1), 113-132.
 DOI: \href{https://doi.org/10.1016/S0168-9002(97)01023-1}{10.1016/S0168-9002(97)01023-1}

 Ryan, C. G., Clayton, E., Griffin, W. L., Sie, S. H. and Cousens, D. R.
 (1988). SNIP, a statistics-sensitive background treatment for the
 quantitative analysis of PIXE spectra in geoscience applications.
 \emph{NIM}, B34 (3), 396-402.
 DOI: \href{https://doi.org/10.1016/0168-583X(88)90063-8}{10.1016/0168-583X(88)90063-8}
}
\author{
N. Frerebeau
}
