% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BDX1}
\alias{BDX1}
\title{CRP2A Calibration Curve #1}
\format{An object of class \code{CalibrationCurve} of length 1.}
\usage{
data("BDX1")
}
\description{
CRP2A Calibration Curve #1
}
\details{
\tabular{ll}{
 \strong{Laboratory:} \tab IRAMAT-CRP2A (UMR 5060) \cr
 \strong{Instrument:} \tab Canberra Inspector 1000 \cr
 \strong{Detector:} \tab LaBr \cr
 \strong{Authors:} \tab CRP2A Luminescence Team
}
}
\examples{
# Load the curve
utils::data("BDX1", package = "gamma")
BDX1

\dontrun{
# See how to replicate the curve
utils::vignette("CRP2A#1", package = "gamma")
}
}
\keyword{datasets}
