% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BDX100_curve}
\alias{BDX100_curve}
\alias{BDX100_data}
\title{CRP2A Calibration Curve LaBr #1}
\format{An object of class \code{CalibrationCurve} of length 1.}
\usage{
data("BDX100_data")
 data("BDX100_curve")

BDX100_data
}
\description{
CRP2A Calibration Curve LaBr #1
}
\details{
\tabular{ll}{
 \strong{Laboratory} \tab IRAMAT-CRP2A (UMR 5060) \cr
 \strong{Instrument} \tab Canberra Inspector 1000 \cr
 \strong{Detector} \tab LaBr \cr
 \strong{Authors} \tab CRP2A Luminescence Team
}
}
\examples{
\dontrun{
## Load the curve
utils::data("BDX100_curve", package = "gamma")
BDX100_curve

## See how to replicate the curve
utils::vignette("CRP2A#1", package = "gamma")
}
}
\seealso{
Other datasets: \code{\link{.decay_La}},
  \code{\link{.decay}}, \code{\link{AIX100_curve}},
  \code{\link{BDX200_curve}}, \code{\link{BDX300_curve}},
  \code{\link{clermont}}
}
\concept{datasets}
\keyword{datasets}
