% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BDX200_curve}
\alias{BDX200_curve}
\alias{BDX200_data}
\title{CRP2A Calibration Curve LaBr #2}
\format{An object of class \code{CalibrationCurve} of length 1.}
\usage{
data("BDX200_data")
 data("BDX200_curve")

BDX200_data
}
\description{
CRP2A Calibration Curve LaBr #2
}
\details{
\tabular{ll}{
 \strong{Laboratory} \tab IRAMAT-CRP2A (UMR 5060) \cr
 \strong{Instrument} \tab Canberra Inspector 1000 \cr
 \strong{Detector} \tab LaBr \cr
 \strong{Authors} \tab CRP2A Luminescence Team
}
}
\examples{
\dontrun{
## Load the curve
utils::data("BDX200_curve", package = "gamma")
BDX200_curve
}
}
\seealso{
Other datasets: \code{\link{.decay_La}},
  \code{\link{.decay}}, \code{\link{AIX100_curve}},
  \code{\link{BDX100_curve}}, \code{\link{BDX300_curve}},
  \code{\link{clermont}}
}
\concept{datasets}
\keyword{datasets}
