% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BDX300}
\alias{BDX300}
\title{CRP2A Calibration Curve NaI}
\format{An object of class \code{CalibrationCurve} of length 1.}
\usage{
data("BDX300")
}
\description{
CRP2A Calibration Curve NaI
}
\details{
\tabular{ll}{
 \strong{Laboratory} \tab IRAMAT-CRP2A (UMR 5060) \cr
 \strong{Instrument} \tab Canberra Inspector 1000 \cr
 \strong{Detector} \tab NaI \cr
 \strong{Authors} \tab CRP2A Luminescence Team
}
}
\examples{
\dontrun{
## Load the curve
utils::data("BDX300", package = "gamma")
BDX300
}
}
\seealso{
Other datasets: \code{\link{.decay_La}},
  \code{\link{.decay}}, \code{\link{AIX100}},
  \code{\link{BDX100}}, \code{\link{BDX200}},
  \code{\link{clermont}}
}
\concept{datasets}
\keyword{datasets}
