% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BDX_LaBr_1}
\alias{BDX_LaBr_1}
\title{CRP2A Calibration Curve (LaBr)}
\format{
An object of class \linkS4class{CalibrationCurve}.
\tabular{ll}{
\strong{Laboratory} \tab IRAMAT-CRP2A (UMR 5060) \cr
\strong{Instrument} \tab Canberra Inspector 1000 \cr
\strong{Detector} \tab LaBr \cr
\strong{Authors} \tab CRP2A Luminescence Team
}
}
\usage{
data(BDX_LaBr_1)
}
\description{
CRP2A Calibration Curve (LaBr)
}
\examples{
## Load the curve
data(BDX_LaBr_1, package = "gamma")
plot(BDX_LaBr_1)
}
\seealso{
Other datasets: 
\code{\link{AIX_NaI_1}},
\code{\link{clermont}},
\code{\link{clermont_2024}},
\code{\link{heidelberg_2024}}
}
\concept{datasets}
\keyword{datasets}
