% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BDX_LaBr_1_data}
\alias{BDX_LaBr_1_data}
\title{CRP2A Calibration Curve Data (LaBr #1)}
\format{
An object of class \code{data.frame} with 7 rows and 10 columns.
}
\usage{
data(BDX_LaBr_1_data)
}
\description{
CRP2A Calibration Curve Data (LaBr #1)
}
\seealso{
\link{BDX_LaBr_1_curve}

Other datasets: 
\code{\link{AIX_NaI_curve}},
\code{\link{AIX_NaI_data}},
\code{\link{BDX_LaBr_1_curve}},
\code{\link{clermont}}
}
\concept{datasets}
\keyword{datasets}
