% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calibrate.R
\docType{methods}
\name{calibrate}
\alias{calibrate}
\alias{calibrate-method}
\alias{calibrate,GammaSpectra-method}
\title{Calibration}
\usage{
calibrate(object, ...)

\S4method{calibrate}{GammaSpectra}(object, dose, peaks = c(238, 1461,
  2614.5), noise = list(value = 1190, error = 1), m = 3, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{GammaSpectra}.}

\item{...}{Currently not used.}

\item{dose}{TODO.}

\item{peaks}{A \code{\link{numeric}} vector.}

\item{noise}{A \code{\link{list}} of numeric values.}

\item{m}{An \code{\link{integer}}.}
}
\description{
Builds a calibration curve.
}
\examples{
# Import CNF files for calibration
dir <- system.file("extdata/calib/", package = "gamma")
calib_spc <- read(dir)

# Build calibration curve
calib_curve <- calibrate(
  calib_spc,
  dose = list(BRIQUE = c(1984.64, 34.08), C347 = c(1421.38, 25.25),
              C341 = c(849.01, 21.32), PEP = c(2535.53, 112.17),
              GOU = c(1573.41, 17.43))
)

# Plot curve
plot(calib_curve)

# Import CNF file
dir <- system.file("extdata/", package = "gamma")
gamma_spc <- read(dir)

adjust(gamma_spc, calib_curve)
}
\author{
N. Frerebeau
}
