% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calibrateEnergy.R
\docType{methods}
\name{calibrateEnergy}
\alias{calibrateEnergy}
\alias{calibrateEnergy-method}
\alias{calibrateEnergy,GammaSpectra-method}
\alias{calibrateEnergy,GammaSpectrum-method}
\title{Spectrum calibration}
\usage{
calibrateEnergy(object, ...)

\S4method{calibrateEnergy}{GammaSpectra}(object, lines = c(238, 1461,
  2614.5), ...)

\S4method{calibrateEnergy}{GammaSpectrum}(object, lines = c(238, 1461,
  2614.5), ...)
}
\arguments{
\item{object}{An object of class \linkS4class{GammaSpectrum}.}

\item{...}{Currently not used.}

\item{lines}{A \code{\link{numeric}} vector giving the peak positions
(in keV) of known sources. The default are 214Pb (238 keV), 40K (1461 keV)
and Cs (2614.5 keV).
Change it only if you know what you are doing.}
}
\value{
An object of class \linkS4class{GammaSpectra}.
}
\description{
Calibrate the energy scale of a gamma spectrum.
}
\details{
TODO
}
\author{
N. Frerebeau
}
