% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CalibrationCurve-class}
\alias{CalibrationCurve-class}
\alias{.CalibrationCurve}
\title{An S4 class to Represent a Dose Rate Calibration Curve}
\description{
An S4 class to Represent a Dose Rate Calibration Curve
}
\section{Slots}{

\describe{
\item{\code{details}}{A \code{\link{list}} of length-one vector giving the curve
metadata.}

\item{\code{model}}{A \code{\link[stats:lm]{linear model}} specifying the calibration
curve.}

\item{\code{noise}}{A length-two \code{\link{numeric}} vector giving the noise value
and error (see \code{\link{integrate_signal}}).}

\item{\code{integration}}{A length-two \code{\link{numeric}} vector giving the energy
range to integrate within (see \code{\link{integrate_signal}}).}

\item{\code{data}}{A \code{\link[=data.frame]{data frame}} giving the data used for
linear model fitting.}
}}

\section{Subset}{

In the code snippets below, \code{x} is a \code{CalibrationCurve} object.
\describe{
 \item{\code{x[[i]]}}{Extracts informations from a slot selected by
 subscript \code{i}. \code{i} is a \code{character} vector of length one.}
}
}

\seealso{
Other class: \code{\link{BaseLine-class}},
  \code{\link{GammaSpectra-class}},
  \code{\link{GammaSpectrum-class}},
  \code{\link{PeakPosition-class}}
}
\author{
N. Frerebeau
}
\concept{class}
