% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/extract.R
\docType{class}
\name{CalibrationCurve-class}
\alias{CalibrationCurve-class}
\alias{[[,CalibrationCurve-method}
\title{An S4 class to represent a calibration curve}
\usage{
\S4method{[[}{CalibrationCurve}(x, i)
}
\arguments{
\item{x}{An object of class \code{CalibrationCurve}.}

\item{i}{A length-one \code{\link{character}} vector specifying the element
to extract or replace (see below). Character sring will be matched to the
names of the slots.}
}
\description{
An S4 class to represent a calibration curve
}
\section{Slots}{

\describe{
\item{\code{instrument}}{A \code{\link{character}} string.}

\item{\code{laboratory}}{A \code{\link{character}} string.}

\item{\code{date}}{A \code{\link[=Date]{date}}.}

\item{\code{model}}{A \code{\link[stats:lm]{linear model}}.}

\item{\code{noise}}{A \code{\link{numeric}} vector.}

\item{\code{data}}{A \code{\link[=data.frame]{data frame}}.}
}}

\section{Subset}{

In the code snippets below, \code{x} is a \code{CalibrationCurve} object.
\describe{
 \item{\code{x[[i]]}}{Extracts informations from a slot selected by
 subscript \code{i}. \code{i} is a \code{character} vector
 of length one.}
}
}

\author{
N. Frerebeau
}
