% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce.R
\docType{methods}
\name{coerce}
\alias{coerce}
\alias{as.matrix.GammaSpectrum}
\title{Coerce}
\usage{
\method{as.matrix}{GammaSpectrum}(x, ...)
}
\arguments{
\item{x}{An object to be coerced.}

\item{...}{Currently not used.}

\item{row.names}{\code{NULL} or a \code{\link{character}} vector giving the
row names for the data frame. Missing values are not allowed.}

\item{optional}{A \code{\link{logical}} scalar (see
\code{\link[base]{as.data.frame}}).}

\item{make.names}{A \code{\link{logical}} scalar (see
\code{\link[base]{as.data.frame}}).}

\item{stringsAsFactors}{A \code{\link{logical}} scalar: should the character
vector be converted to a \code{\link{factor}}?}
}
\value{
A coerced object.
}
\description{
Coerce
}
\examples{
## Import a Canberra CNF file
spc_file <- system.file("extdata/LaBr.CNF", package = "gamma")
spc <- read(spc_file)

## Coerce
mtx <- as.matrix(spc)
df <- as.data.frame(spc)
head(df)
}
\author{
N. Frerebeau
}
\concept{class}
