% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{decay}
\alias{decay}
\title{Nuclear Decay Data}
\format{
A \code{\link{data.frame}} with the following columns (absolute errors):
\describe{
\item{decay_chain}{}
\item{isotope}{}
\item{occurrence}{}
\item{occurrence_error}{}
\item{post_radon}{}
\item{half_life}{}
\item{half_life_error}{}
\item{energy}{Energy in keV.}
\item{energy_error}{}
\item{gamma_intensity}{}
\item{gamma_intensity_error}{}
\item{counts_chain}{}
\item{counts_chain_error}{}
}
}
\source{
Nuclides And Isotopes : Chart of the Nuclides.
}
\description{
An internal dataset containing the decay data for the following isotopes:
Th-232, U-235, U-238 and K-40.
}
\keyword{datasets}
\keyword{internal}
