% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{.decay_La}
\alias{.decay_La}
\title{Lanthanum Decay Data}
\format{
A \eqn{2094 \times 11}{2094 x 11} \code{\link[=data.frame]{data frame}}
 with the following columns (absolute errors):
 \describe{
   \item{decay_chain}{}
   \item{isotope}{}
   \item{occurrence}{}
   \item{occurrence_error}{}
   \item{post_radon}{}
   \item{half_life}{}
   \item{half_life_error}{}
   \item{energy}{Energy in keV.}
   \item{energy_error}{}
   \item{gamma_intensity}{}
   \item{gamma_intensity_error}{}
 }
}
\source{
\href{IAEA Live Chart of Nuclides}{https://www-nds.iaea.org/relnsd/vcharthtml/VChartHTML.html}.
 Accessed 2019-06-07.
}
\usage{
.decay_La
}
\description{
An internal dataset containing the decay data for the Lanthanum isotopes.
}
\keyword{datasets}
\keyword{internal}
