% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/doserate.R
\docType{methods}
\name{doserate}
\alias{doserate}
\alias{fit_dose}
\alias{fit_dose-method}
\alias{predict_dose}
\alias{predict_dose-method}
\alias{fit_dose,GammaSpectra-method}
\alias{predict_dose,CalibrationCurve,GammaSpectrum-method}
\alias{predict_dose,CalibrationCurve,GammaSpectra-method}
\title{Dose Rate Estimation}
\usage{
fit_dose(object, ...)

predict_dose(object, spectrum, ...)

\S4method{fit_dose}{GammaSpectra}(object, Ni_noise, Ni_range, NiEi_noise,
  NiEi_range, details = NULL, ...)

\S4method{predict_dose}{CalibrationCurve,GammaSpectrum}(object, spectrum,
  threshold = c("Ni", "NiEi"), epsilon = 0, ...)

\S4method{predict_dose}{CalibrationCurve,GammaSpectra}(object, spectrum,
  threshold = c("Ni", "NiEi"), epsilon = 0, ...)
}
\arguments{
\item{object}{A \linkS4class{GammaSpectra} or \linkS4class{CalibrationCurve}
object.}

\item{...}{Currently not used.}

\item{spectrum}{An optional \linkS4class{GammaSpectrum} or
\linkS4class{GammaSpectra} object in which to look for variables with which
to predict. If omitted, the fitted values are used.}

\item{Ni_noise, NiEi_noise}{A length-two \code{\link{numeric}} vector giving
the background noise integration value and error, respectively.}

\item{Ni_range, NiEi_range}{A length-two \code{\link{numeric}} vector giving
the energy range to integrate within (in keV).}

\item{details}{A \code{\link{list}} of length-one vector specifying
additional informations about the instrument for which the curve is built.}

\item{threshold}{A \code{\link{character}} sting specifying the threshold
to be used. It must be one of \code{"Ni"} (default) or \code{"NiEi"}
(see details).}

\item{epsilon}{A \code{\link{numeric}} value giving an extra error term
introduced by the calibration of the energy scale of the spectrum.}
}
\value{
A \linkS4class{CalibrationCurve} object.
}
\description{
\code{fit_dose} builds a calibration curve for gamma dose rate estimation.
}
\details{
\code{predict_dose} predicts in-situ gamma dose rate.
}
\examples{
# Import CNF files for calibration
spc_dir <- system.file("extdata/BDX100/calibration", package = "gamma")
spectra <- read(spc_dir)

# Set dose rate values and errors for each spectrum
data("clermont")
set_dose(spectra) <- clermont[, c("gamma_dose", "gamma_error")]

# Build the calibration curve
calib_curve <- fit_dose(
  spectra,
  Ni_noise = c(22.61, 0.05),
  Ni_range = c(300, 2800),
  NiEi_noise = c(25279.63, 1.66),
  NiEi_range = c(165, 2800)
)

# Check the linear model
summary(get_model(calib_curve, "Ni"))
summary(get_model(calib_curve, "NiEi"))

# Plot the curve
plot(calib_curve, threshold = "Ni")

# Estimate gamma dose rates
predict_dose(calib_curve, spectra, threshold = "Ni")
predict_dose(calib_curve, spectra, threshold = "NiEi")
}
\references{
Mercier, N. & Falguères, C. (2007). Field Gamma Dose-Rate Measurement with
 a NaI(Tl) Detector: Re-Evaluation of the "Threshold" Technique.
 \emph{Ancient TL}, 25(1), p. 1-4.
}
\seealso{
\link{integrate_signal}
}
\author{
N. Frerebeau
}
\concept{dose rate}
