% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/estimateDoseRate.R
\docType{methods}
\name{estimateDoseRate}
\alias{estimateDoseRate}
\alias{estimateDoseRate-method}
\alias{estimateDoseRate,GammaSpectrum,CalibrationCurve-method}
\alias{estimateDoseRate,GammaSpectrum-method}
\alias{estimateDoseRate,GammaSpectra,CalibrationCurve-method}
\alias{estimateDoseRate,GammaSpectra-method}
\title{Gamma dose rate}
\usage{
estimateDoseRate(object, curve, ...)

\S4method{estimateDoseRate}{GammaSpectrum,CalibrationCurve}(object, curve,
  ...)

\S4method{estimateDoseRate}{GammaSpectra,CalibrationCurve}(object, curve,
  epsilon = 0.03, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{GammaSpectra}.}

\item{curve}{An object of class \linkS4class{CalibrationCurve}.}

\item{...}{Extra parameters passed to \code{\link{integrateSignal}}.}

\item{epsilon}{A \code{\link{numeric}} value.}
}
\description{
Estimates in-situ gamma dose rate.
}
\examples{
# Import CNF files for calibration
dir <- system.file("extdata/cerege/", package = "gamma")
calib_spc <- read(dir, skip = 1:35)[-5]

# Calibrate energy scale
calib_lines <- list(
  Pb = c(chanel = 76, energy = 238),
  K = c(chanel = 459, energy = 1461),
  Cs = c(chanel = 816, energy = 2614.5)
)

calib_scale <- calibrateEnergy(calib_spc, lines = calib_lines)

# Build calibration curve
known_dose <- list(
  BRIQUE = c(1984.64, 34.08),
  C347 = c(1421.38, 25.25),
  C341 = c(849.01, 21.32),
  PEP = c(2535.53, 112.17),
  GOU = c(1573.41, 17.43)
)

calib_curve <- calibrateDose(
  calib_spc,
  dose = known_dose,
  noise = c(value = 1190, error = 1),
  range = c(200, 2800)
)

# Plot curve
plot(calib_curve) +
  ggplot2::labs(x = "Dose rate [µGy/y]", y = "Integrated signal") +
  ggplot2::theme_bw()

# Import CNF file
dir <- system.file("extdata/", package = "gamma")
gamma_spc <- read(dir, skip = 1:35)

# Estimate gamma dose rate
dose_rate <- estimateDoseRate(gamma_spc, calib_curve)
}
\author{
N. Frerebeau
}
