% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/extract.R
\docType{methods}
\name{extract}
\alias{extract}
\alias{extract-method}
\alias{replace-method}
\alias{getDoseRate}
\alias{setDoseRate<-}
\alias{getDoseRate,GammaSpectrum-method}
\alias{setDoseRate<-,GammaSpectrum-method}
\alias{setDoseRate,GammaSpectrum-method}
\alias{getDoseRate,GammaSpectra-method}
\alias{setDoseRate<-,GammaSpectra-method}
\alias{setDoseRate,GammaSpectra-method}
\title{Extract or Replace Parts of an Object}
\usage{
getDoseRate(object)

setDoseRate(object) <- value

\S4method{getDoseRate}{GammaSpectrum}(object)

\S4method{setDoseRate}{GammaSpectrum}(object) <- value

\S4method{getDoseRate}{GammaSpectra}(object)

\S4method{setDoseRate}{GammaSpectra}(object) <- value
}
\arguments{
\item{object}{An object from which to extract element(s) or in which to
replace element(s).}

\item{value}{A possible value for the element(s) of \code{object}}
}
\description{
Extract or Replace Parts of an Object
}
\author{
N. Frerebeau
}
