% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/findPeaks.R
\docType{methods}
\name{findPeaks}
\alias{findPeaks}
\alias{findPeaks-method}
\alias{findPeaks,GammaSpectra-method}
\alias{findPeaks,GammaSpectrum-method}
\title{Find peaks}
\usage{
findPeaks(object, ...)

\S4method{findPeaks}{GammaSpectra}(object, span = NULL, ...)

\S4method{findPeaks}{GammaSpectrum}(object, method = c("MAD"), SNR = 2,
  span = NULL, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{GammaSpectrum}.}

\item{...}{Currently not used.}

\item{span}{An \code{\link{integer}} giving the half window size. If
\code{NULL}, 5\% of the number of chanels is used as the half window size.}

\item{method}{A \code{\link{character}} string.}

\item{SNR}{An \code{\link{integer}}.}
}
\value{
A three columns data frame or a list of data frame.
}
\description{
Finds local maxima in sequential data.
}
\details{
A local maximum has to be the highest one in the given window and has to be
 higher than \eqn{SNR \times noise}{SNR * noise} to be recognized as peak.
}
\note{
Adapted from Stasia Grinberg's \href{https://github.com/stas-g/findPeaks}{algorithm}.
}
\examples{
# Import a CNF file
spc_dir <- system.file("extdata/test1.cnf", package = "gamma")
spc <- read(spc_dir)

# Estimate baseline
baseline <- estimateBaseline(spc)

# Plot spectrum + baseline
plot(spc, baseline) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts")

# Remove baseline
spc_clean <- removeBaseline(spc)

# Detect peaks
peak_index <- findPeaks(spc_clean, span = 20)

plot(spc_clean, select = 1:3, facet = TRUE) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts") +
  ggplot2::geom_vline(xintercept = peak_index$energy, colour = "red") +
  ggplot2::theme_bw()
}
\author{
N. Frerebeau
}
