% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/fit.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit-method}
\alias{fit,GammaSpectra,DoseRate,numeric-method}
\title{Calibration}
\usage{
fit(object, doses, noise, ...)

\S4method{fit}{GammaSpectra,DoseRate,numeric}(object, doses, noise,
  range = c(200, 2800), intercept = TRUE, weights = FALSE,
  details = NULL, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{GammaSpectra}.}

\item{doses}{A \code{\link{list}} of length-two numeric vectors giving the
dose rate values and errors, respectively, for each element of
\code{object}. Elements of the list will be matched by names to the
references of the spectra.}

\item{noise}{A \code{\link{list}} of two numeric values giving the noise
value and error.}

\item{...}{Currently not used.}

\item{range}{A length-two \code{\link{numeric}} vector giving the energy
range to integrate within (in keV).}

\item{intercept}{A \code{\link{logical}} scalar: should the intercept of the
curve be estimated?}

\item{weights}{A \code{\link{logical}} scalar: should weights be used in the
fitting process. If \code{TRUE}, the inverse of the squared dose errors are
used.}

\item{details}{A list of \code{\link{character}} vector specifying additional
informations about the instrument for which the curve is built.}
}
\value{
An object of class \linkS4class{CalibrationCurve}.
}
\description{
Builds a calibration curve for gamma dose rate estimation.
}
\examples{
# Import CNF files for calibration
spc_dir <- system.file("extdata/crp2a/", package = "gamma")
spc <- read(spc_dir, skip = TRUE)
## Select 'BRIQUE', 'C341', 'C347', 'GOU', 'LMP', 'MAZ' and 'PEP' to build the curve
spc_calib <- spc[c("BRIQUE", "C341", "C347", "GOU", "LMP", "MAZ", "PEP")]

# Set dose rate values and errors for each spectrum
known_doses <- list(
  reference = c("BRIQUE", "C341", "C347", "GOU", "LMP", "MAZ", "PEP"),
  dose_value = c(1986, 850, 1424, 1575, 642, 1141, 2538),
  dose_error = c(36, 21, 24, 17, 18, 12, 112)
)
## Coerce to a 'DoseRate' object
known_doses <- methods::as(known_doses, "DoseRate")

# Build the calibration curve
calib_curve <- fit(
  spc_calib,
  doses = known_doses,
  noise = c(value = 1190, error = 1),
  range = c(200, 2800)
)

# Check the linear model
summary(calib_curve[["model"]])

# Plot the curve
plot(calib_curve) +
  ggplot2::labs(x = "Signal", y = "Dose rate [µGy/y]")

# Estimate gamma dose rates
(dose_rate <- predict(calib_curve, spc_calib))

# Plot the estimated values
plot(calib_curve, dose_rate) +
  ggplot2::labs(x = "Signal", y = "Dose rate [µGy/y]")
}
\seealso{
\link{predict}, \link{integrateSignal}
}
\author{
N. Frerebeau
}
