% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{FWHM}
\alias{FWHM}
\title{FWHM}
\usage{
FWHM(x, y, center)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector giving the \eqn{x} and \eqn{y}
coordinates of a set of points. Alternatively, a single argument \eqn{x}
can be provided.}

\item{center}{A \code{\link{numeric}} value giving the peak position in
\code{x} units.}
}
\value{
A \code{\link{numeric}} value.
}
\description{
Estimates the Half-Width at Half-Maximum (FWHM) for a given peak.
}
\details{
It tries to get the smallest possible estimate.
}
\author{
N. Frerebeau
}
\keyword{internal}
