% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\docType{package}
\name{gamma-package}
\alias{gamma-package}
\alias{_PACKAGE}
\alias{gamma}
\title{gamma: Dose Rate Estimation from In-Situ Gamma-Ray Spectrometry Measurements for
    Luminescence Dating}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

'gamma' is intended to process in-situ
    gamma-ray spectrometry measurements for luminescence dating. This
    package allows to import, inspect and (automatically) correct the energy 
    shifts of the spectrum. It provides methods for estimating the gamma 
    dose rate by the use of a calibration curve. The package only supports 
    Canberra CNF and TKA files.
}
\details{
\tabular{ll}{
 \strong{Package:} \tab gamma \cr
 \strong{Type:} \tab Package \cr
 \strong{Version:} \tab 0.0.0.9000 \cr
 \strong{Date:} \tab 2019-03-28 \cr
 \strong{License:} \tab GPL-3 \cr
 \strong{DOI:} \tab \href{https://doi.org/xxx}{xxx}
}
}
\note{
This work received a state financial support managed by the Agence Nationale
de la Recherche (France) throught the program \emph{Investissements d'avenir}
(ref. ANR-10-LABX-52).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://gamma.archaeo.science}
  \item \url{http://github.com/crp2a/gamma}
  \item Report bugs at \url{http://github.com/crp2a/gamma/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
 Nicolas Frerebeau \tab \emph{IRAMAT-CRP2A, Université Bordeaux Montaigne, France} \cr
 Guillaume Guérin \tab \emph{IRAMAT-CRP2A, Université Bordeaux Montaigne, France} \cr
 Christelle Lahaye \tab \emph{IRAMAT-CRP2A, Université Bordeaux Montaigne, France} \cr
 Brice Lebrun \tab \emph{CEREGE, Aix-Marseille Université, France}
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Universite Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
