% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GammaSpectra-class}
\alias{GammaSpectra-class}
\alias{.GammaSpectra}
\title{An S4 Class to Represent a Collection of Gamma Sectra}
\arguments{
\item{x}{An object of class \code{GammaSpectra}.}
}
\description{
Represents a collection of spectra of gamma ray spectrometry measurements.
}
\details{
This class extends the base \code{\link{list}} and can only contains
 \linkS4class{GammaSpectrum} objects.
}
\section{Access}{

In the code snippets below, \code{x} is a \code{GammaSpectra} object.
\describe{
 \item{\code{length(x)}}{Get the number of elements in \code{x}.}
 \item{\code{names(x)}}{Get the names of the elements of \code{x}.}
 \item{\code{get_hash(x)}}{Get the MD5 hash of the raw data file.}
 \item{\code{get_chanels(x)}}{Get the number of chanels of \code{x}.}
 \item{\code{get_energy(x)}}{Get the energy scales of \code{x}.}
 \item{\code{get_dose(x)}}{Get the dose rates of \code{x}.}
}
}

\section{Subset}{

In the code snippets below, \code{x} is a \code{GammaSpectra} object.
\describe{
 \item{\code{x[i]}}{Extracts the elements selected by subscript \code{i}.
  \code{i} can be \code{missing} or \code{NULL}, \code{numeric} or
  \code{character} vector or a \code{factor}.
  Returns a new \code{GammaSpectra} object.}
 \item{\code{x[i, j]}}{Like the above but allows to select a slot thru
  \code{j} (see examples). \code{j} is a \code{character} vector of
  length one. Returns a \code{list}.}
 \item{\code{x[[i]]}}{Extracts the elements selected by subscript \code{i}.
  \code{i} can be a \code{numeric} or \code{character} vector
  of length one. Returns the corresponding \linkS4class{GammaSpectrum}
  object.}
}
}

\examples{
# Import all CNF files in a given directory
spc_dir <- system.file("extdata/crp2a/calibration", package = "gamma")
(spectra <- read(spc_dir))

# Access
names(spectra)
length(spectra)

# Coerce
as(spectra, "list")

# Subset
spectra[] # All spectra
spectra[NULL] # All spectra
spectra[1] # The first spectrum
spectra[-6] # Delete the sixth spectrum
spectra[1:3] # The first three spectra
spectra[c(1, 3)] # The first and third spectra
spectra["BRIQUE"] # The spectrum named 'BRIQUE'
spectra[c("BRIQUE", "C347")] # The spectra named 'BRIQUE' and 'C347'
spectra[1:3, "energy"] # The slot 'energy' of the first three spectra
spectra[[1]]
spectra[["BRIQUE"]]
}
\seealso{
\linkS4class{GammaSpectrum}.

Other class: \code{\link{BaseLine-class}},
  \code{\link{CalibrationCurve-class}},
  \code{\link{GammaSpectrum-class}},
  \code{\link{PeakPosition-class}}
}
\author{
N. Frerebeau
}
\concept{class}
