% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/extract.R
\docType{class}
\name{GammaSpectra-class}
\alias{GammaSpectra-class}
\alias{[,GammaSpectra-method}
\alias{[[,GammaSpectra-method}
\title{An S4 class to represent a collection of gamma sectra}
\usage{
\S4method{[}{GammaSpectra}(x, i, j)

\S4method{[[}{GammaSpectra}(x, i)
}
\arguments{
\item{x}{An object of class \code{GammaSpectrum}.}

\item{i, j}{Indices specifying elements to extract or replace (see below).
Indices are \code{\link{numeric}} or \code{\link{character}} vectors or
empty (\code{\link{missing}}) or \code{\link{NULL}}.
Numeric values are coerced to integer as by \code{\link{as.integer}}
(and hence truncated towards zero). Character vectors will be matched to
the names of the object.}
}
\description{
An S4 class to represent a collection of gamma sectra
}
\section{Access}{

In the code snippets below, \code{x} is a \code{GammaSpectra} object.
\describe{
 \item{\code{length(x)}}{Get the number of elements in \code{x}.}
 \item{\code{names(x)}, \code{names(x) <- value}}{Get or set the names of the
  elements according to \code{value}.}
}
}

\section{Coerce}{

In the code snippets below, \code{x} is a \code{GammaSpectra} object.
\describe{
 \item{\code{as(x, "list")}}{Coerces \code{x} to a \code{\link{list}}.}
}
}

\section{Subset}{

In the code snippets below, \code{x} is a \code{GammaSpectra} object.
\describe{
 \item{\code{x[i]}}{Extracts the elements selected by subscript \code{i}.
  \code{i} can be \code{missing} or \code{NULL}, \code{numeric} or
  \code{character} vector or a \code{factor}.
  Returns a new \code{GammaSpectra} object.}
 \item{\code{x[i, j]}}{Like the above but allows to select a slot thru
  \code{j} (see examples). \code{j} is a \code{character} vector of
  length one. Returns a \code{list}.}
 \item{\code{x[[i]]}}{Extracts the elements selected by subscript \code{i}.
  \code{i} can be a \code{numeric} or \code{character} vector
  of length one. Returns the corresponding \linkS4class{GammaSpectrum} object.}
}
}

\examples{
# Import all CNF files in a given directory
dir <- system.file("extdata/calib/", package = "gamma")
(spectra <- read(dir))

# Access
names(spectra)
length(spectra)

# Coerce
as(spectra, "list")

# Subset
spectra[] # All spectra
spectra[NULL] # All spectra
spectra[1] # The first spectrum
spectra[-6] # Delete the sixth spectrum
spectra[1:3] # The first three spectra
spectra[c(1, 3)] # The first and third spectra
spectra["BRIQUE"] # The spectrum named 'BRIQUE'
spectra[c("BRIQUE", "C347")] # The spectra named 'BRIQUE' and 'C347'
spectra[1:3, "energy"] # The slot 'energy' of the first three spectra
spectra[[1]]
spectra[["BRIQUE"]]
}
\author{
N. Frerebeau
}
