% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/extract.R
\docType{class}
\name{GammaSpectrum-class}
\alias{GammaSpectrum-class}
\alias{.GammaSpectrum}
\alias{[[,GammaSpectrum-method}
\alias{length,GammaSpectrum-method}
\title{An S4 class to represent a gamma sectrum}
\usage{
\S4method{[[}{GammaSpectrum}(x, i)

\S4method{length}{GammaSpectrum}(x)
}
\arguments{
\item{x}{An object of class \code{GammaSpectrum}.}

\item{i}{A length-one \code{\link{character}} vector specifying the element
to extract or replace (see below). Character sring will be matched to the
names of the slots.}
}
\value{
TODO
}
\description{
Represents a single spectrum of a gamma ray spectrometry measurement.
}
\section{Slots}{

\describe{
\item{\code{hash}}{A \code{\link{character}} string giving the 32-byte MD5 hash of
the imported file.}

\item{\code{reference}}{A \code{\link{character}} string the measurement reference.}

\item{\code{date}}{A \code{\link{POSIXct}} element giving the measurement date and
time.}

\item{\code{instrument}}{A \code{\link{character}} string giving the instrument name.}

\item{\code{file_format}}{A \code{\link{character}} string giving the format of the
imported file.}

\item{\code{live_time}}{A \code{\link{numeric}} value.}

\item{\code{real_time}}{A \code{\link{numeric}} value.}

\item{\code{chanel}}{A \code{\link{integer}} vector giving the channel number.
Numeric values are coerced to integer as by \code{\link{as.integer}}
(and hence truncated towards zero).}

\item{\code{energy}}{A \code{\link{numeric}} vector giving the gamma ray's energy
(in keV).}

\item{\code{counts}}{A \code{\link{numeric}} vector giving the counts number for
each channel. Numeric values are coerced to integer as by
\code{\link{as.integer}} (and hence truncated towards zero).}

\item{\code{rate}}{A \code{\link{numeric}} vector the count rate (in 1/s) for
each channel.}

\item{\code{calibration}}{A \code{\link[stats:lm]{linear model}} used for energy
scale calibration (see \code{\link{calibrate}}).}

\item{\code{dose_rate}}{A length-two \code{\link{numeric}} vector giving the dose
rate and corresponding error.}
}}

\section{Access}{

In the code snippets below, \code{x} is a \code{GammaSpectrum} object.
\describe{
 \item{\code{length(x)}}{Get the number of chanels in \code{x}.}
}
}

\section{Coerce}{

In the code snippets below, \code{x} is a \code{GammaSpectrum} object.
\describe{
 \item{\code{as(x, "matrix")}}{Coerces \code{x} to a \code{\link{matrix}}.}
 \item{\code{as(x, "data.frame")}}{Coerces \code{x} to a
 \code{\link[=data.frame]{data frame}}.}
}
}

\section{Subset}{

In the code snippets below, \code{x} is a \code{GammaSpectrum} object.
\describe{
 \item{\code{x[[i]]}}{Extracts informations from a slot selected by
 subscript \code{i}. \code{i} is a \code{character} vector
 of length one.}
}
}

\examples{
# Import a Canberra CNF file
spc_file <- system.file("extdata/test_CNF.cnf", package = "gamma")
(spectrum <- read(spc_file))

# Coerce
df <- as(spectrum, "data.frame")
head(df)
tail(df)

# Subset
spectrum[["hash"]]
spectrum[["reference"]]
spectrum[["date"]]
spectrum[["instrument"]]
spectrum[["file_format"]]
spectrum[["live_time"]]
spectrum[["real_time"]]
spectrum[["calibration"]]
}
\seealso{
\linkS4class{GammaSpectra}, \linkS4class{BaseLine}
}
\author{
N. Frerebeau
}
