% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/extract.R
\docType{class}
\name{GammaSpectrum-class}
\alias{GammaSpectrum-class}
\alias{[[,GammaSpectrum-method}
\title{An S4 class to represent a gamma sectrum}
\usage{
\S4method{[[}{GammaSpectrum}(x, i)
}
\arguments{
\item{x}{An object of class \code{GammaSpectrum}.}

\item{i}{A length-one \code{\link{character}} vector specifying the element
to extract or replace (see below). Character sring will be matched to the
names of the slots.}
}
\description{
An S4 class to represent a gamma sectrum
}
\section{Slots}{

\describe{
\item{\code{reference}}{A \code{\link{character}} string the measurement reference.}

\item{\code{date}}{A \code{\link{character}} string giving the measurement date.}

\item{\code{instrument}}{A \code{\link{character}} string the instrument name.}

\item{\code{file_format}}{A \code{\link{character}} string.}

\item{\code{live_time}}{A \code{\link{numeric}} value.}

\item{\code{real_time}}{A \code{\link{numeric}} value.}

\item{\code{chanel}}{A \code{\link{numeric}} vector.}

\item{\code{energy}}{A \code{\link{numeric}} vector.}

\item{\code{counts}}{A \code{\link{numeric}} vector.}
}}

\section{Coerce}{

In the code snippets below, \code{x} is a \code{GammaSpectrum} object.
\describe{
 \item{\code{as(x, "matrix")}}{Coerces \code{x} to a \code{\link{matrix}}.}
 \item{\code{as(x, "data.frame")}}{Coerces \code{x} to a
 \code{\link[=data.frame]{data frame}}.}
}
}

\section{Subset}{

In the code snippets below, \code{x} is a \code{GammaSpectrum} object.
\describe{
 \item{\code{x[[i]]}}{Extracts informations from a slot selected by
 subscript \code{i}. \code{i} is a \code{character} vector
 of length one.}
}
}

\examples{
# Import a Canberra CNF file
file <- system.file("extdata/test1.cnf", package = "gamma")
(spectrum <- read(file)[[1]])

# Coerce
as(spectrum, "data.frame")

# Subset
spectrum[["reference"]]
spectrum[["instrument"]]
}
\author{
N. Frerebeau
}
