% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HD_CZT_1}
\alias{HD_CZT_1}
\title{heiLUM Calibration Curve (CZT)}
\format{
Objects of class \linkS4class{CalibrationCurve}.
\tabular{ll}{
\strong{Laboratory} \tab heiLUM, Heidelberg University, DE \cr
\strong{Instrument} \tab Kromek GR1 & Kromek RayMon10 (SN: 5261)\cr
\strong{Detector} \tab GR1+ (SN: 1625) & GR1 (SN: 1663) \cr
\strong{Authors} \tab Sebastian Kreutzer, Didier Maillier, Norbert Mercier, heiLUM Team
}

The calibration dataset consists of the following objects:
\itemize{
\item CAL_GR1_AdamiecAitken1998
\item CAL_GR1_Cresswelletal2018
\item CAL_GR1_Gurinetal2011
\item CAL_GR1_Liritzisetal2013
\item CAL_RAYMON_GR1_AdamiecAitken1998
\item CAL_RAYMON_GR1_Cresswelletal2018
\item CAL_RAYMON_GR1_Gurinetal2011
\item CAL_RAYMON_GR1_Liritzisetal2013
}
}
\usage{
data(HD_CZT_1)
}
\description{
heiLUM Calibration Curve (CZT)
}
\examples{
## Load the curve
data(HD_CZT_1, package = "gamma")
names(HD_CZT_1)
plot(HD_CZT_1$CAL_GR1_AdamiecAitken1998)

}
\references{
Kreutzer, S., Martin, L., Miallier, D. & Mercier, N. (2025).
Environmental gamma dose rate measurements using cadmium zinc telluride (CZT) detectors.
\emph{Geochronology}, 7, p. 229–246. \doi{10.5194/gchron-7-229-2025}

Kreutzer, S., Martin, L., Miallier, D., Mercier, N. & Faske, M. (2024).
Dataset: Environmental Gamma Dose Rate Measurements using CZT Detectors (1.0.0).
\emph{Zenodo}. \doi{10.5281/zenodo.13731839}
}
\seealso{
Other datasets: 
\code{\link{AIX_NaI_1}},
\code{\link{BDX_LaBr_1}},
\code{\link{clermont}},
\code{\link{clermont_2024}},
\code{\link{heidelberg_2024}}
}
\concept{datasets}
\keyword{datasets}
