% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Shiny App}
\usage{
launch_app()
}
\value{
shiny application object
}
\description{
A wrapper for shiny::shinyApp()
}
\examples{
\dontrun{launch_app()}
}
\author{
N. Frerebeau
}
