% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{LLS}
\alias{LLS}
\alias{inverseLLS}
\title{LLS operator}
\usage{
LLS(x)

inverseLLS(x)
}
\arguments{
\item{x}{A \code{numeric} vector.}
}
\value{
A \code{numeric} vector.
}
\description{
LLS operator
}
\references{
Morháč, M., Kliman, J., Matoušek, V., Veselský, M. and Turzo, I. (1997).
 Background elimination methods for multidimensional gamma-ray spectra.
 \emph{Nuclear Instruments and Methods in Physics Research Section A:
 Accelerators, Spectrometers, Detectors and Associated Equipment}, 401(1), p. 113-132.
 DOI: \href{https://doi.org/10.1016/S0168-9002(97)01023-1}{10.1016/S0168-9002(97)01023-1}

 Morháč, M. and Matoušek, V. (2008). Peak Clipping Algorithms for Background
 Estimation in Spectroscopic Data. \emph{Applied Spectroscopy}, 62(1), p. 91-106.
 DOI: \href{https://doi.org/10.1366/000370208783412762}{10.1366/000370208783412762}
}
\author{
N. Frerebeau
}
\keyword{internal}
