% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{MAD}
\alias{MAD}
\title{MAD}
\usage{
MAD(x, k = 1.4826, na.rm = FALSE)
}
\arguments{
\item{x}{A \code{\link{numeric}} vector.}

\item{k}{A \code{\link{numeric}} value.}

\item{na.rm}{A \code{\link{logical}} scalar.}
}
\value{
A \code{numeric} value.
}
\description{
Calculates the Median Absolute Deviation (MAD).
}
\author{
N. Frerebeau
}
\keyword{internal}
