% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{mutator}
\alias{mutator}
\alias{get}
\alias{set}
\alias{get_hash}
\alias{get_hash-method}
\alias{get_names}
\alias{get_names-method}
\alias{set_names<-}
\alias{set_names-method}
\alias{get_livetime}
\alias{get_livetime-method}
\alias{get_realtime}
\alias{get_realtime-method}
\alias{get_chanels}
\alias{get_chanels-method}
\alias{get_counts}
\alias{get_counts-method}
\alias{get_rates}
\alias{get_rates-method}
\alias{get_energy}
\alias{get_energy-method}
\alias{set_energy<-}
\alias{set_energy-method}
\alias{get_model}
\alias{get_model-method}
\alias{get_noise}
\alias{get_noise-method}
\alias{get_range}
\alias{get_range-method}
\alias{range_chanels}
\alias{range_chanels-method}
\alias{range_energy}
\alias{range_energy-method}
\alias{length,GammaSpectrum-method}
\alias{get_hash,GammaSpectrum-method}
\alias{get_names,GammaSpectrum-method}
\alias{get_livetime,GammaSpectrum-method}
\alias{get_realtime,GammaSpectrum-method}
\alias{get_chanels,GammaSpectrum-method}
\alias{get_counts,GammaSpectrum-method}
\alias{get_rates,GammaSpectrum-method}
\alias{get_energy,GammaSpectrum-method}
\alias{range_energy,GammaSpectrum-method}
\alias{range_chanels,GammaSpectrum-method}
\alias{get_hash,GammaSpectra-method}
\alias{get_names,GammaSpectra-method}
\alias{get_livetime,GammaSpectra-method}
\alias{get_realtime,GammaSpectra-method}
\alias{get_chanels,GammaSpectra-method}
\alias{get_counts,GammaSpectra-method}
\alias{get_rates,GammaSpectra-method}
\alias{get_energy,GammaSpectra-method}
\alias{range_energy,GammaSpectra-method}
\alias{range_chanels,GammaSpectra-method}
\alias{get_model,CalibrationCurve-method}
\alias{get_noise,CalibrationCurve-method}
\alias{get_range,CalibrationCurve-method}
\alias{get_hash,PeakPosition-method}
\alias{get_chanels,PeakPosition-method}
\alias{get_energy,PeakPosition-method}
\alias{set_names<-,GammaSpectrum-method}
\alias{set_names,GammaSpectrum-method}
\alias{set_names<-,GammaSpectra-method}
\alias{set_name,GammaSpectra-method}
\alias{set_energy<-,PeakPosition,numeric-method}
\alias{set_energy,PeakPosition,numeric-method}
\title{Get or Set Parts of an Object}
\usage{
get_hash(x)

get_names(x)

set_names(x) <- value

get_livetime(x)

get_realtime(x)

get_chanels(x)

get_counts(x)

get_rates(x)

get_energy(x)

set_energy(x) <- value

get_model(x, ...)

get_noise(x, ...)

get_range(x, ...)

range_chanels(x, ...)

range_energy(x, ...)

\S4method{length}{GammaSpectrum}(x)

\S4method{get_hash}{GammaSpectrum}(x)

\S4method{get_names}{GammaSpectrum}(x)

\S4method{get_livetime}{GammaSpectrum}(x)

\S4method{get_realtime}{GammaSpectrum}(x)

\S4method{get_chanels}{GammaSpectrum}(x)

\S4method{get_counts}{GammaSpectrum}(x)

\S4method{get_rates}{GammaSpectrum}(x)

\S4method{get_energy}{GammaSpectrum}(x)

\S4method{range_energy}{GammaSpectrum}(x, na.rm = FALSE)

\S4method{range_chanels}{GammaSpectrum}(x, na.rm = FALSE)

\S4method{get_hash}{GammaSpectra}(x)

\S4method{get_names}{GammaSpectra}(x)

\S4method{get_livetime}{GammaSpectra}(x)

\S4method{get_realtime}{GammaSpectra}(x)

\S4method{get_chanels}{GammaSpectra}(x)

\S4method{get_counts}{GammaSpectra}(x)

\S4method{get_rates}{GammaSpectra}(x)

\S4method{get_energy}{GammaSpectra}(x)

\S4method{range_energy}{GammaSpectra}(x, na.rm = FALSE)

\S4method{range_chanels}{GammaSpectra}(x, na.rm = FALSE)

\S4method{get_model}{CalibrationCurve}(x, threshold = c("Ni", "NiEi"))

\S4method{get_noise}{CalibrationCurve}(x, threshold = c("Ni", "NiEi"))

\S4method{get_range}{CalibrationCurve}(x, threshold = c("Ni", "NiEi"))

\S4method{get_hash}{PeakPosition}(x)

\S4method{get_chanels}{PeakPosition}(x)

\S4method{get_energy}{PeakPosition}(x)

\S4method{set_names}{GammaSpectrum}(x) <- value

\S4method{set_names}{GammaSpectra}(x) <- value

\S4method{set_energy}{PeakPosition,numeric}(x) <- value
}
\arguments{
\item{x}{An object from which to get or set element(s).}

\item{value}{A possible value for the element(s) of \code{x}.}

\item{...}{Currently not used.}

\item{na.rm}{A \code{\link{logical}} scalar: should \code{\link{NA}} be
omitted?}

\item{threshold}{A \code{\link{character}} sting specifying the threshold
to be used. It must be one of \code{"Ni"} (default) or \code{"NiEi"}
(see details).}

\item{simplify}{A \code{\link{logical}} scalar: should the result be
simplified to a matrix? If \code{FALSE} (default), returns a list.}
}
\value{
An object of the same sort as \code{x} with the new values assigned.
}
\description{
Getters and setters to extract or replace parts of an object.
}
\seealso{
Other mutator: 
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutator}
