% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\docType{methods}
\name{operators}
\alias{operators}
\alias{operators-method}
\alias{-,GammaSpectrum,GammaSpectrum-method}
\alias{-,GammaSpectrum,BaseLine-method}
\title{Operators}
\usage{
\S4method{-}{GammaSpectrum,GammaSpectrum}(e1, e2)

\S4method{-}{GammaSpectrum,BaseLine}(e1, e2)
}
\arguments{
\item{e1, e2}{Objects (typically \linkS4class{GammaSpectrum} objects).}
}
\value{
A \linkS4class{GammaSpectrum} object.
}
\description{
Performs operation on spectra.
}
\author{
N. Frerebeau
}
