% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/extract.R
\docType{class}
\name{PeakPosition-class}
\alias{PeakPosition-class}
\alias{.PeakPosition}
\alias{[[,PeakPosition-method}
\title{An S4 class to represent a set of peaks}
\usage{
\S4method{[[}{PeakPosition}(x, i)
}
\arguments{
\item{x}{An object of class \code{PeakPosition}.}

\item{i}{A length-one \code{\link{character}} vector specifying the element
to extract or replace (see below). Character sring will be matched to the
names of the slots.}
}
\value{
TODO
}
\description{
An S4 class to represent a set of peaks
}
\section{Slots}{

\describe{
\item{\code{hash}}{TODO.}

\item{\code{noise_method}}{A \code{\link{character}} string specifying the method
used for peak detection.}

\item{\code{noise_threshold}}{A length one \code{\link{numeric}} vector giving the
noise threshold.}

\item{\code{window}}{A length one \code{\link{numeric}} vector giving the half-window
size.}

\item{\code{chanel}}{TODO.}

\item{\code{energy}}{TODO.}
}}

\section{Subset}{

In the code snippets below, \code{x} is a \code{PeakPosition} object.
\describe{
 \item{\code{x[[i]]}}{Extracts informations from a slot selected by
 subscript \code{i}. \code{i} is a \code{character} vector
 of length one.}
}
}

\author{
N. Frerebeau
}
