% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/peaks.R
\docType{methods}
\name{peaks}
\alias{peaks}
\alias{findPeaks}
\alias{findPeaks-method}
\alias{fitPeaks}
\alias{fitPeaks-method}
\alias{findPeaks,GammaSpectra-method}
\alias{findPeaks,GammaSpectrum-method}
\alias{fitPeaks,PeakPosition-method}
\alias{fitPeaks,GammaSpectrum-method}
\title{Peaks}
\usage{
findPeaks(object, ...)

fitPeaks(object, ...)

\S4method{findPeaks}{GammaSpectra}(object, span = NULL, ...)

\S4method{findPeaks}{GammaSpectrum}(object, method = c("MAD"), SNR = 2,
  span = NULL, ...)

\S4method{fitPeaks}{PeakPosition}(object, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{GammaSpectrum}.}

\item{...}{Currently not used.}

\item{span}{An \code{\link{integer}} giving the half window size. If
\code{NULL}, 5\% of the number of chanels is used as the half window size.}

\item{method}{A \code{\link{character}} string.}

\item{SNR}{An \code{\link{integer}}.}
}
\value{
\code{findPeaks} returns an object of class \linkS4class{PeakPosition}.

 \code{fitPeaks} returns an object of class \linkS4class{PeakModel}.
}
\description{
Finds local maxima in sequential data.
}
\details{
A local maximum has to be the highest one in the given window and has to be
 higher than \eqn{SNR \times noise}{SNR * noise} to be recognized as peak.
}
\examples{
# Import a CNF file
spc_dir <- system.file("extdata/test1.cnf", package = "gamma")
spc <- read(spc_dir)

# Estimate baseline
baseline <- estimateBaseline(spc)

# Plot spectrum + baseline
plot(spc, baseline) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts")

# Remove baseline
spc_clean <- removeBaseline(spc)

# Detect peaks
peak_index <- findPeaks(spc_clean, span = 20)

plot(peak_index) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts") +
  ggplot2::theme_bw()
}
\author{
N. Frerebeau
}
