% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/predicates.R
\docType{methods}
\name{predicates}
\alias{predicates}
\alias{is_calibrated}
\alias{is_calibrated,GammaSpectrum-method}
\alias{is_calibrated,GammaSpectra-method}
\title{Predicates}
\usage{
is_calibrated(object)

\S4method{is_calibrated}{GammaSpectrum}(object)

\S4method{is_calibrated}{GammaSpectra}(object)
}
\arguments{
\item{object}{An object from which to get or set element(s).}
}
\value{
A \code{\link{logical}} vector.
}
\description{
Predicates
}
\author{
N. Frerebeau
}
\concept{predicates}
