% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/predict.R
\docType{methods}
\name{predict}
\alias{predict}
\alias{predict-method}
\alias{predict,CalibrationCurve-method}
\title{Gamma dose rate}
\usage{
\S4method{predict}{CalibrationCurve}(object, spectra, epsilon = 0.03,
  simplify = FALSE, ...)
}
\arguments{
\item{object}{A \linkS4class{CalibrationCurve} object.}

\item{spectra}{An optional \linkS4class{GammaSpectra} object in
which to look for variables with which to predict. If omitted, the fitted
values are used.}

\item{epsilon}{A \code{\link{numeric}} value giving the error introduced by
the energy calibration of the spectrum.}

\item{simplify}{A \code{\link{logical}} scalar: should the result be
simplified to a matrix? If \code{FALSE} (default), returns a list.}

\item{...}{Currently not used.}
}
\value{
If \code{simplify} is \code{FALSE} returns a list of length-two numeric
 vectors (default), else returns a matrix.
}
\description{
Predict in-situ gamma dose rate.
}
\examples{
# Import CNF files for calibration
spc_dir <- system.file("extdata/crp2a/calibration", package = "gamma")
spc_calib <- read(spc_dir, skip = TRUE)

# Set dose rate values and errors for each spectrum
data("clermont")
setDoseRate(spc_calib) <- clermont[, c("gamma", "gamma_error")]

# Build the calibration curve
calib_curve <- fit(
  spc_calib,
  noise = c(value = 1190, error = 1),
  range = c(200, 2800)
)

# Check the linear model
summary(calib_curve[["model"]])

# Plot the curve
plot(calib_curve) +
  ggplot2::labs(x = "Signal", y = "Dose rate [µGy/y]")

# Estimate gamma dose rates
(dose_rate <- predict(calib_curve, spc_calib))
}
\seealso{
Other dose rate: \code{\link{fit}}
}
\author{
N. Frerebeau
}
\concept{dose rate}
