% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/baseline.R
\docType{methods}
\name{processBaseline}
\alias{processBaseline}
\alias{estimateBaseline}
\alias{estimateBaseline-method}
\alias{removeBaseline}
\alias{removeBaseline-method}
\alias{estimateBaseline,GammaSpectra-method}
\alias{estimateBaseline,GammaSpectrum-method}
\alias{removeBaseline,GammaSpectra-method}
\alias{removeBaseline,GammaSpectrum-method}
\title{Baseline estimation and removal}
\usage{
estimateBaseline(object, ...)

removeBaseline(object, ...)

\S4method{estimateBaseline}{GammaSpectra}(object, method = c("SNIP"),
  ...)

\S4method{estimateBaseline}{GammaSpectrum}(object, method = c("SNIP"),
  ...)

\S4method{removeBaseline}{GammaSpectra}(object, method = c("SNIP"), ...)

\S4method{removeBaseline}{GammaSpectrum}(object, method = c("SNIP"), ...)
}
\arguments{
\item{object}{A \code{\link{numeric}} vector.}

\item{...}{Extra parameters passed to internal functions.}

\item{method}{A \code{\link{character}} string specifying the method to be
used for baseline estimation.}
}
\value{
\code{estimateBaseline} and \code{removeBaseline} return a
 \linkS4class{GammaSpectrum} or \linkS4class{GammaSpectra} object (same as
 \code{object}).
}
\description{
Baseline estimation and removal
}
\examples{
# Import a CNF file
spc_file <- system.file("extdata/test_CNF.cnf", package = "gamma")
spectrum <- read(spc_file, skip = TRUE)

# Estimate baseline
baseline <- estimateBaseline(spectrum)

plot(spectrum, baseline) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts")

# Remove baseline
spc_clean1 <- removeBaseline(spectrum)
spc_clean2 <- spectrum - baseline

plot(spc_clean1) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts")
}
\references{
Morháč, M., Kliman, J., Matoušek, V., Veselský, M. and Turzo, I. (1997).
 Background elimination methods for multidimensional gamma-ray spectra.
 \emph{Nuclear Instruments and Methods in Physics Research Section A:
 Accelerators, Spectrometers, Detectors and Associated Equipment}, 401(1), p. 113-132.
 DOI: \href{https://doi.org/10.1016/S0168-9002(97)01023-1}{10.1016/S0168-9002(97)01023-1}

 Morháč, M. and Matoušek, V. (2008). Peak Clipping Algorithms for Background
 Estimation in Spectroscopic Data. \emph{Applied Spectroscopy}, 62(1), p. 91-106.
 DOI: \href{https://doi.org/10.1366/000370208783412762}{10.1366/000370208783412762}

 Ryan, C. G., Clayton, E., Griffin, W. L., Sie, S. H. and Cousens, D. R.
 (1988). SNIP, a statistics-sensitive background treatment for the
 quantitative analysis of PIXE spectra in geoscience applications.
 \emph{Nuclear Instruments and Methods in Physics Research Section B:
 Beam Interactions with Materials and Atoms}, 34(3), p. 396-402.
 DOI: \href{https://doi.org/10.1016/0168-583X(88)90063-8}{10.1016/0168-583X(88)90063-8}
}
\author{
N. Frerebeau
}
