% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{readCSV}
\alias{readCSV}
\title{Temporary import method for csv files}
\usage{
readCSV(file, ...)
}
\arguments{
\item{file}{The name of the file which the data are to be read from.}

\item{...}{Extra parameters passed to \code{\link[utils]{read.table}}}
}
\value{
A \linkS4class{GammaSpectrum} object.
}
\description{
Temporary import method for csv files
}
\examples{
csv <- system.file("extdata/T14.csv", package = "gamma")

readCSV(csv, dec = ".")
}
