% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/simulate-methods.R
\docType{methods}
\name{simulate}
\alias{simulate}
\alias{simulate-method}
\alias{simulate,numeric,numeric,numeric-method}
\title{Simulate a Gamma Spectrum}
\usage{
simulate(K, Th, U, ...)

\S4method{simulate}{numeric,numeric,numeric}(K, Th, U, energy = c(1,
  3200), n = 1024, ...)
}
\arguments{
\item{K}{TODO}

\item{Th}{TODO}

\item{U}{TODO}

\item{...}{Currently not used.}

\item{energy}{A length-two \code{\link{numeric}} vector giving the energy
range (in keV).}

\item{n}{An \code{\link{integer}} giving TODO.}
}
\value{
A \linkS4class{GammaSpectrum} object.
}
\description{
Simulate a Gamma Spectrum
}
\author{
N. Frerebeau
}
