% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/signal.R
\docType{methods}
\name{slice_signal}
\alias{slice_signal}
\alias{slice_signal-method}
\alias{slice_signal,GammaSpectrum-method}
\alias{slice_signal,GammaSpectra-method}
\title{Choose Chanels by Position}
\usage{
slice_signal(object, ...)

\S4method{slice_signal}{GammaSpectrum}(object, ...)

\S4method{slice_signal}{GammaSpectra}(object, ...)

\S4method{slice_signal}{GammaSpectra}(object, ...)
}
\arguments{
\item{object}{A \linkS4class{GammaSpectrum} or \linkS4class{GammaSpectra}
object.}

\item{...}{\code{\link{integer}} values giving the chanels of the
spectrum to be kept/dropped (see below). Numeric values are coerced to
integer as by \code{\link{as.integer}} (and hence truncated towards zero).}
}
\value{
A \linkS4class{GammaSpectrum} or \linkS4class{GammaSpectra} object.
}
\description{
Choose chanels by position.
}
\details{
Either positive values to keep, or negative values to drop, should be
 provided. The values provided must be either all positive or all negative.

 If no value is provided, an attempt is made to define the number
 of channels to skip at the beginning of the spectrum. This drops all
 channels before the highest count maximum. This is intended to deal with the
 artefact produced by the rapid growth of random background noise towards low
 energies.
}
\examples{
## Import CNF files
spc_file <- system.file("extdata/test_LaBr.CNF", package = "gamma")
spectrum <- read(spc_file)

## Plot spectrum
plot(spectrum)

## Slice
sliced <- slice_signal(spectrum, 450:550)
plot(sliced)

sliced <- slice_signal(spectrum)
plot(sliced)
}
\seealso{
Other signal processing: 
\code{\link{baseline}},
\code{\link{find_peaks}()},
\code{\link{integrate_signal}()},
\code{\link{smooth_signal}()},
\code{\link{stabilize_signal}()}
}
\author{
N. Frerebeau
}
\concept{signal processing}
