% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/smooth.R
\docType{methods}
\name{smooth}
\alias{smooth}
\alias{smooth-method}
\alias{smooth,GammaSpectra-method}
\alias{smooth,GammaSpectrum-method}
\title{Smooth}
\usage{
smooth(object, ...)

\S4method{smooth}{GammaSpectra}(object, method = c("unweighted",
  "weighted"), m = 3, ...)

\S4method{smooth}{GammaSpectrum}(object, method = c("unweighted",
  "weighted"), m = 3, ...)
}
\arguments{
\item{object}{A spectrum to be smoothed (a \linkS4class{GammaSpectrum} or
\linkS4class{GammaSpectra} object).}

\item{...}{Currently not used.}

\item{method}{A \code{\link{character}} string specifying the smoothing
method to be used. It must be one of "\code{unweighted}" (default) or
"\code{weighted}". Any unambiguous substring can be given.}

\item{m}{An odd \code{\link{integer}} value giving the number of adjacent
points to mean.}
}
\value{
A \linkS4class{GammaSpectrum} or \linkS4class{GammaSpectra} object.
}
\description{
Smooth
}
\details{
The following smoothing methods are available:
 \describe{
  \item{unweighted}{Unweighted sliding-average or rectangular smooth.
  It replaces each point in the signal with the average of \code{m} adjacent
  points.}
  \item{weighted}{Weighted sliding-average or triangular smooth.
  It replaces each point in the signal with the weighted mean of \code{m}
  adjacent points.}
 }
 There will be \eqn{(m - 1)/2} points both at the beginning and at the end of
 the spectrum for which a complete \code{m}-width smooth cannot be
 calculated. To prevent data loss, progressively smaller smooths are used at
 the ends of the spectra.
}
\seealso{
Other signal processing: \code{\link{integrateSignal}},
  \code{\link{peaks}}, \code{\link{processBaseline}}
}
\author{
N. Frerebeau
}
\concept{signal processing}
