% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smoothSavitzky}
\alias{smoothSavitzky}
\alias{coefficientSavitzky}
\title{Savitzky-Golay Filter}
\usage{
smoothSavitzky(x, m, p = 2)

coefficientSavitzky(m, p = 2)
}
\arguments{
\item{x}{A \code{\link{numeric vector}} of observed values to be smoothed.}

\item{m}{An odd \code{\link{integer}} scalar giving the number of adjacent
points to use.}

\item{p}{An \code{\link{integer}} scalar giving the polynomial degree.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
\code{smoothSavitzky} smoothes the data using the Savitzky-Golay filter.
\code{coefficientSavitzky} computes the Savitzky-Golay convolution
coefficients.
}
\keyword{internal}
