% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smoothTriangular}
\alias{smoothTriangular}
\title{Triangular Smooth}
\usage{
smoothTriangular(x, m)
}
\arguments{
\item{x}{A \code{\link{numeric vector}} of observed values to be smoothed.}

\item{m}{An odd \code{\link{integer}} scalar giving the number of adjacent
points to use.}
}
\value{
A \code{\link{numeric}} vector of the same length as \code{x}.
}
\description{
Weighted sliding-average.
}
\keyword{internal}
