% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/signal.R
\docType{methods}
\name{stabilize_signal}
\alias{stabilize_signal}
\alias{stabilize_signal-method}
\alias{stabilize_signal,GammaSpectrum-method}
\alias{stabilize_signal,GammaSpectra-method}
\title{Transform Intensities}
\usage{
stabilize_signal(object, ...)

\S4method{stabilize_signal}{GammaSpectrum}(object, transformation, ...)

\S4method{stabilize_signal}{GammaSpectra}(object, transformation, ...)
}
\arguments{
\item{object}{A \linkS4class{GammaSpectrum} object.}

\item{...}{Extra arguments to be passed to \code{transformation}.}

\item{transformation}{A \code{\link{function}} that takes a numeric vector as
argument and returns a numeric vector.}
}
\value{
A new \linkS4class{GammaSpectrum} object with transformed
 intensities.
}
\description{
Transform Intensities
}
\seealso{
Other signal processing: \code{\link{baseline}},
  \code{\link{find_peaks}}, \code{\link{integrate_signal}},
  \code{\link{slice_signal}}, \code{\link{smooth_signal}}
}
\author{
N. Frerebeau
}
\concept{signal processing}
