% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/show.R
\docType{methods}
\name{summarise}
\alias{summarise}
\alias{summarise-method}
\alias{summarise,GammaSpectrum-method}
\alias{summarise,GammaSpectra-method}
\title{Summarize}
\usage{
summarise(object, ...)

\S4method{summarise}{GammaSpectrum}(object)

\S4method{summarise}{GammaSpectra}(object)
}
\arguments{
\item{object}{A \linkS4class{GammaSpectrum} or \linkS4class{GammaSpectra}
object.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{data.frame}}.
}
\description{
Summarize
}
\examples{
## Import a Canberra CNF file
cnf_file <- system.file("extdata/test_LaBr.CNF", package = "gamma")
(spectrum <- read(cnf_file))

summarise(spectrum)

## Import all CNF files in a given directory
spc_dir <- system.file("extdata/BDX100/calibration", package = "gamma")
(spectra <- read(spc_dir))

summarise(spectra)
}
\seealso{
Other IO: \code{\link{operators}}, \code{\link{plot}},
  \code{\link{read}}
}
\author{
N. Frerebeau
}
\concept{IO}
