% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/extract.R
\docType{methods}
\name{access}
\alias{access}
\alias{get}
\alias{set}
\alias{get_hash}
\alias{get_hash-method}
\alias{get_names}
\alias{get_names-method}
\alias{set_names<-}
\alias{set_names-method}
\alias{get_time}
\alias{get_time-method}
\alias{get_nchanels}
\alias{get_nchanels-method}
\alias{get_chanels}
\alias{get_chanels-method}
\alias{get_counts}
\alias{get_counts-method}
\alias{get_rates}
\alias{get_rates-method}
\alias{get_energy}
\alias{get_energy-method}
\alias{set_energy<-}
\alias{set_energy-method}
\alias{get_dose}
\alias{get_dose-method}
\alias{set_dose<-}
\alias{set_dose-method}
\alias{get_model}
\alias{get_model-method}
\alias{get_noise}
\alias{get_noise-method}
\alias{get_range}
\alias{get_range-method}
\alias{range_energy}
\alias{range_energy-method}
\alias{get_hash,GammaSpectrum-method}
\alias{get_names,GammaSpectrum-method}
\alias{set_names<-,GammaSpectrum-method}
\alias{set_names,GammaSpectrum-method}
\alias{get_time,GammaSpectrum-method}
\alias{get_chanels,GammaSpectrum-method}
\alias{get_nchanels,GammaSpectrum-method}
\alias{get_counts,GammaSpectrum-method}
\alias{get_rates,GammaSpectrum-method}
\alias{get_energy,GammaSpectrum-method}
\alias{get_dose,GammaSpectrum-method}
\alias{set_dose<-,GammaSpectrum-method}
\alias{set_dose,GammaSpectrum-method}
\alias{range_energy,GammaSpectrum-method}
\alias{get_hash,GammaSpectra-method}
\alias{get_names,GammaSpectra-method}
\alias{set_names<-,GammaSpectra-method}
\alias{set_name,GammaSpectra-method}
\alias{get_time,GammaSpectra-method}
\alias{get_chanels,GammaSpectra-method}
\alias{get_nchanels,GammaSpectra-method}
\alias{get_counts,GammaSpectra-method}
\alias{get_rates,GammaSpectra-method}
\alias{get_energy,GammaSpectra-method}
\alias{get_dose,GammaSpectra-method}
\alias{set_dose<-,GammaSpectra-method}
\alias{set_dose,GammaSpectra-method}
\alias{range_energy,GammaSpectra-method}
\alias{get_model,CalibrationCurve-method}
\alias{get_noise,CalibrationCurve-method}
\alias{get_range,CalibrationCurve-method}
\alias{get_hash,PeakPosition-method}
\alias{get_chanels,PeakPosition-method}
\alias{get_energy,PeakPosition-method}
\alias{set_energy<-,PeakPosition-method}
\alias{set_energy,PeakPosition-method}
\title{Get or Set Parts of an Object}
\usage{
get_hash(object)

get_names(object)

set_names(object) <- value

get_time(object, ...)

get_nchanels(object)

get_chanels(object, ...)

get_counts(object, ...)

get_rates(object, ...)

get_energy(object, ...)

set_energy(object) <- value

get_dose(object)

set_dose(object) <- value

get_model(object, ...)

get_noise(object, ...)

get_range(object, ...)

range_energy(object)

\S4method{get_hash}{GammaSpectrum}(object)

\S4method{get_names}{GammaSpectrum}(object)

\S4method{set_names}{GammaSpectrum}(object) <- value

\S4method{get_time}{GammaSpectrum}(object, type = c("live", "real"))

\S4method{get_chanels}{GammaSpectrum}(object)

\S4method{get_nchanels}{GammaSpectrum}(object)

\S4method{get_counts}{GammaSpectrum}(object)

\S4method{get_rates}{GammaSpectrum}(object)

\S4method{get_energy}{GammaSpectrum}(object)

\S4method{get_dose}{GammaSpectrum}(object)

\S4method{set_dose}{GammaSpectrum}(object) <- value

\S4method{range_energy}{GammaSpectrum}(object)

\S4method{get_hash}{GammaSpectra}(object)

\S4method{get_names}{GammaSpectra}(object)

\S4method{set_names}{GammaSpectra}(object) <- value

\S4method{get_time}{GammaSpectra}(object, type = c("live", "real"))

\S4method{get_chanels}{GammaSpectra}(object, simplify = FALSE)

\S4method{get_nchanels}{GammaSpectra}(object)

\S4method{get_counts}{GammaSpectra}(object, simplify = FALSE)

\S4method{get_rates}{GammaSpectra}(object, simplify = FALSE)

\S4method{get_energy}{GammaSpectra}(object, simplify = FALSE)

\S4method{get_dose}{GammaSpectra}(object)

\S4method{set_dose}{GammaSpectra}(object) <- value

\S4method{range_energy}{GammaSpectra}(object)

\S4method{get_model}{CalibrationCurve}(object, threshold = c("Ni", "NiEi"))

\S4method{get_noise}{CalibrationCurve}(object, threshold = c("Ni", "NiEi"))

\S4method{get_range}{CalibrationCurve}(object, threshold = c("Ni", "NiEi"))

\S4method{get_hash}{PeakPosition}(object)

\S4method{get_chanels}{PeakPosition}(object)

\S4method{get_energy}{PeakPosition}(object)

\S4method{set_energy}{PeakPosition}(object) <- value
}
\arguments{
\item{object}{An object from which to get or set element(s).}

\item{value}{A possible value for the element(s) of \code{object} (see
below).}

\item{...}{Currently not used.}

\item{type}{A \code{\link{character}} string specifying the value to be
returned. It must be one of "\code{live}" (the default) or "\code{real}".}

\item{simplify}{A \code{\link{logical}} scalar: should the result be
simplified to a matrix? If \code{FALSE} (default), returns a list.}

\item{threshold}{A \code{\link{character}} sting specifying the threshold
to be used. It must be one of \code{"Ni"} (default) or \code{"NiEi"}
(see details).}
}
\value{
An object of the same sort as \code{object} with the new values assigned.
}
\description{
Getters and setters to extract or replace parts of an object.
}
\seealso{
Other mutator: 
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutator}
