% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/extract.R
\docType{methods}
\name{access}
\alias{access}
\alias{get}
\alias{set}
\alias{get_hash}
\alias{get_reference}
\alias{set_reference<-}
\alias{get_chanels}
\alias{get_dose}
\alias{set_dose<-}
\alias{get_energy}
\alias{set_energy<-}
\alias{get_hash,GammaSpectrum-method}
\alias{get_reference,GammaSpectrum-method}
\alias{set_reference<-,GammaSpectrum-method}
\alias{set_reference,GammaSpectrum-method}
\alias{get_chanels,GammaSpectrum-method}
\alias{get_length,GammaSpectrum-method}
\alias{get_energy,GammaSpectrum-method}
\alias{get_dose,GammaSpectrum-method}
\alias{set_dose<-,GammaSpectrum-method}
\alias{set_dose,GammaSpectrum-method}
\alias{get_hash,GammaSpectra-method}
\alias{get_reference,GammaSpectra-method}
\alias{set_reference<-,GammaSpectra-method}
\alias{set_reference,GammaSpectra-method}
\alias{get_chanels,GammaSpectra-method}
\alias{get_length,GammaSpectra-method}
\alias{get_energy,GammaSpectra-method}
\alias{get_dose,GammaSpectra-method}
\alias{set_dose<-,GammaSpectra-method}
\alias{set_dose,GammaSpectra-method}
\alias{get_chanels,PeakPosition-method}
\alias{get_energy,PeakPosition-method}
\alias{set_energy<-,PeakPosition-method}
\alias{set_energy,PeakPosition-method}
\title{Get or Set Parts of an Object}
\usage{
get_hash(object)

get_reference(object)

set_reference(object) <- value

get_chanels(object)

get_dose(object)

set_dose(object) <- value

get_energy(object)

set_energy(object) <- value

\S4method{get_hash}{GammaSpectrum}(object)

\S4method{get_reference}{GammaSpectrum}(object)

\S4method{set_reference}{GammaSpectrum}(object) <- value

\S4method{get_chanels}{GammaSpectrum}(object)

\S4method{get_energy}{GammaSpectrum}(object)

\S4method{get_dose}{GammaSpectrum}(object)

\S4method{set_dose}{GammaSpectrum}(object) <- value

\S4method{get_hash}{GammaSpectra}(object)

\S4method{get_reference}{GammaSpectra}(object)

\S4method{set_reference}{GammaSpectra}(object) <- value

\S4method{get_chanels}{GammaSpectra}(object)

\S4method{get_energy}{GammaSpectra}(object)

\S4method{get_dose}{GammaSpectra}(object)

\S4method{set_dose}{GammaSpectra}(object) <- value

\S4method{get_chanels}{PeakPosition}(object)

\S4method{get_energy}{PeakPosition}(object)

\S4method{set_energy}{PeakPosition}(object) <- value
}
\arguments{
\item{object}{An object from which to get or set element(s).}

\item{value}{A possible value for the element(s) of \code{object} (see
below).}
}
\value{
An object of the same sort as \code{object} with the new values assigned.
}
\description{
Getters and setters to extract or replace parts of an object.
}
\seealso{
Other mutator: \code{\link{subset}}
}
\author{
N. Frerebeau
}
\concept{mutator}
