% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calibrate.R
\docType{methods}
\name{adjust}
\alias{adjust}
\alias{adjust-method}
\alias{adjust,GammaSpectra,CalibrationCurve-method}
\alias{adjust,GammaSpectra-method}
\title{Gamma dose rate}
\usage{
adjust(object, curve, ...)

\S4method{adjust}{GammaSpectra,CalibrationCurve}(object, curve,
  noise = NULL, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{GammaSpectra}.}

\item{curve}{An object of class \linkS4class{CalibrationCurve}.}

\item{...}{Extra parameters passed to \code{\link{integrateSignal}}.}

\item{noise}{A \code{\link{list}} of numeric values.}
}
\description{
Estimates gamma dose rate.
}
\examples{
# Import CNF files for calibration
dir <- system.file("extdata/cerege/", package = "gamma")
calib_spc <- read(dir)[-5]

# Build calibration curve
calib_curve <- calibrate(
  calib_spc,
  dose = list(BRIQUE = c(1984.64, 34.08), C347 = c(1421.38, 25.25),
              C341 = c(849.01, 21.32), PEP = c(2535.53, 112.17),
              GOU = c(1573.41, 17.43))
)

# Plot curve
plot(calib_curve) +
  ggplot2::labs(x = "Dose rate [µGy/y]", y = "Integrated signal") +
  ggplot2::theme_bw()

# Import CNF file
dir <- system.file("extdata/", package = "gamma")
gamma_spc <- read(dir)

dose_rate <- adjust(gamma_spc, calib_curve)
as(dose_rate, "data.frame")

#plot(calib_curve, dose_rate)
}
\author{
N. Frerebeau
}
