% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BaseLine-class}
\alias{BaseLine-class}
\alias{.BaseLine}
\title{An S4 Class to Represent a Spectrum Baseline}
\description{
An S4 Class to Represent a Spectrum Baseline
}
\note{
This class extends the \linkS4class{GammaSpectrum} class.
}
\examples{
# Import a CNF file
spc_file <- system.file("extdata/test_LaBr.CNF", package = "gamma")
spectrum <- read(spc_file)

# Remove the first chanels
spectrum <- slice_signal(spectrum)

# Estimate baseline
baseline <- estimate_baseline(spectrum)

plot(spectrum, baseline) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts")

# Remove baseline
spc_clean1 <- remove_baseline(spectrum)
spc_clean2 <- spectrum - baseline

plot(spc_clean1) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts")
}
\seealso{
\linkS4class{GammaSpectrum}.

Other class: 
\code{\link{DoseRateModel-class}},
\code{\link{GammaSpectra-class}},
\code{\link{GammaSpectrum-class}},
\code{\link{PeakPosition-class}}
}
\author{
N. Frerebeau
}
\concept{class}
