% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BaseLine-class}
\alias{BaseLine-class}
\title{An S4 class to represent a spectrum baseline}
\description{
An S4 class to represent a spectrum baseline
}
\seealso{
\linkS4class{GammaSpectrum}.
}
\author{
N. Frerebeau
}
