% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/operators.R
\docType{class}
\name{BaseLine-class}
\alias{BaseLine-class}
\alias{-,GammaSpectrum,BaseLine-method}
\title{An S4 class to represent a spectrum baseline}
\usage{
\S4method{-}{GammaSpectrum,BaseLine}(e1, e2)
}
\arguments{
\item{e1, e2}{XXX.}
}
\description{
An S4 class to represent a spectrum baseline
}
\examples{
# Import a CNF file
spc_dir <- system.file("extdata/test1.cnf", package = "gamma")
spc_dir <- system.file("extdata/cerege", package = "gamma")
spc <- read(spc_dir)[[5]]

# Estimate baseline
baseline <- estimateBaseline(spc)

# Plot spectrum + baseline
plot(spc, baseline) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts")

# Remove baseline
spc_clean <- removeBaseline(spc)

# Detect peaks
peak_index <- findPeaks(spc_clean, span = NULL)

plot(spc_clean, select = 1:3, facet = TRUE) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts") +
  ggplot2::geom_vline(xintercept = peak_index$energy, colour = "red") +
  ggplot2::theme_bw()
}
\seealso{
\linkS4class{GammaSpectrum}.
}
\author{
N. Frerebeau
}
