% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BaseLine-class}
\alias{BaseLine-class}
\alias{.BaseLine}
\title{An S4 class to represent a spectrum baseline}
\description{
An S4 class to represent a spectrum baseline
}
\note{
This class extends the \linkS4class{GammaSpectrum} class.
}
\examples{
# Import a CNF file
spc_file <- system.file("extdata/test_CNF.cnf", package = "gamma")
spectrum <- read(spc_file, skip = TRUE)

# Estimate baseline
baseline <- estimateBaseline(spectrum)

plot(spectrum, baseline) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts")

# Remove baseline
spc_clean1 <- removeBaseline(spectrum)
spc_clean2 <- spectrum - baseline

plot(spc_clean1) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts")
}
\seealso{
\linkS4class{GammaSpectrum}.
}
\author{
N. Frerebeau
}
