% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/baseline.R
\docType{methods}
\name{baseline}
\alias{baseline}
\alias{estimate_baseline}
\alias{estimate_baseline-method}
\alias{remove_baseline}
\alias{remove_baseline-method}
\alias{estimate_baseline,GammaSpectra-method}
\alias{estimate_baseline,GammaSpectrum-method}
\alias{remove_baseline,GammaSpectra-method}
\alias{remove_baseline,GammaSpectrum-method}
\title{Baseline Estimation and Removal}
\usage{
estimate_baseline(object, ...)

remove_baseline(object, ...)

\S4method{estimate_baseline}{GammaSpectra}(object, method = c("SNIP"),
  LLS = FALSE, decreasing = FALSE, k = 100)

\S4method{estimate_baseline}{GammaSpectrum}(object, method = c("SNIP"),
  LLS = FALSE, decreasing = FALSE, k = 100)

\S4method{remove_baseline}{GammaSpectra}(object, method = c("SNIP"), ...)

\S4method{remove_baseline}{GammaSpectrum}(object, method = c("SNIP"),
  ...)
}
\arguments{
\item{object}{A \linkS4class{GammaSpectrum} or \linkS4class{GammaSpectra}
object.}

\item{...}{Extra parameters passed to \code{estimate_baseline}.}

\item{method}{A \code{\link{character}} string specifying the method to be
used for baseline estimation (see details).}

\item{LLS}{A \code{\link{logical}} scalar: should the LLS operator be applied
on \code{x} before employing SNIP algorithm? Only used if
\code{method} is "\code{SNIP}".}

\item{decreasing}{A \code{\link{logical}} scalar: should a decreasing
clipping window be used? Only used if \code{method} is "\code{SNIP}".}

\item{k}{An \code{\link{integer}} value giving the numerber of iterations.
Only used if \code{method} is "\code{SNIP}".}
}
\value{
\code{estimate_baseline} returns a \linkS4class{BaseLine} object.

 \code{remove_baseline} returns a \linkS4class{GammaSpectrum} or
 \linkS4class{GammaSpectra} object (same as \code{object}).
}
\description{
Baseline Estimation and Removal
}
\details{
The following methods are availble for baseline estimation:
 \describe{
  \item{SNIP}{Sensitive Nonlinear Iterative Peak clipping algorithm.}
 }
}
\examples{
# Import a CNF file
spc_file <- system.file("extdata/test_CNF.cnf", package = "gamma")
spectrum <- read(spc_file)

# Remove the first chanels
spectrum <- slice_signal(spectrum)

# Estimate baseline
baseline <- estimate_baseline(spectrum)

plot(spectrum, baseline) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts")

# Remove baseline
spc_clean1 <- remove_baseline(spectrum)
spc_clean2 <- spectrum - baseline

plot(spc_clean1) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts")
}
\references{
Morháč, M., Kliman, J., Matoušek, V., Veselský, M. & Turzo, I. (1997).
 Background elimination methods for multidimensional gamma-ray spectra.
 \emph{Nuclear Instruments and Methods in Physics Research Section A:
 Accelerators, Spectrometers, Detectors and Associated Equipment}, 401(1),
 p. 113-132.
 DOI: \href{https://doi.org/10.1016/S0168-9002(97)01023-1}{10.1016/S0168-9002(97)01023-1}

 Morháč, M. & Matoušek, V. (2008). Peak Clipping Algorithms for Background
 Estimation in Spectroscopic Data. \emph{Applied Spectroscopy}, 62(1), p. 91-106.
 DOI: \href{https://doi.org/10.1366/000370208783412762}{10.1366/000370208783412762}

 Ryan, C. G., Clayton, E., Griffin, W. L., Sie, S. H. & Cousens, D. R.
 (1988). SNIP, a statistics-sensitive background treatment for the
 quantitative analysis of PIXE spectra in geoscience applications.
 \emph{Nuclear Instruments and Methods in Physics Research Section B:
 Beam Interactions with Materials and Atoms}, 34(3), p. 396-402.
 DOI: \href{https://doi.org/10.1016/0168-583X(88)90063-8}{10.1016/0168-583X(88)90063-8}
}
\seealso{
Other signal processing: \code{\link{find_peaks}},
  \code{\link{integrate_signal}},
  \code{\link{slice_signal}}, \code{\link{smooth_signal}},
  \code{\link{stabilize_signal}}
}
\author{
N. Frerebeau
}
\concept{signal processing}
