% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calibrate.R
\docType{methods}
\name{calibrate_energy}
\alias{calibrate_energy}
\alias{calibrate_energy-method}
\alias{calibrate_energy,GammaSpectrum,list-method}
\alias{calibrate_energy,GammaSpectrum,PeakPosition-method}
\title{Spectrum calibration}
\usage{
calibrate_energy(object, lines, ...)

\S4method{calibrate_energy}{GammaSpectrum,list}(object, lines, ...)

\S4method{calibrate_energy}{GammaSpectrum,PeakPosition}(object, lines, ...)
}
\arguments{
\item{object}{A \linkS4class{GammaSpectrum} or
\linkS4class{GammaSpectra} object.}

\item{lines}{A list of or a \code{\link{numeric}} vector.
If a \code{list} is provided, each element must be a named length-two
numeric vector giving the observed peak position (chanel) and the
corresponding expected energy value (in keV). If \code{lines} is a
numeric \code{vector}, each element must be an expected energy value.}

\item{...}{Currently not used.}
}
\value{
A \linkS4class{GammaSpectrum} object.
}
\description{
Calibrate the energy scale of a gamma spectrum.
}
\details{
TODO
}
\examples{
## Import a CNF file for calibration
spc_file <- system.file("extdata/test_TKA.tka", package = "gamma")
(spectrum <- read(spc_file))

# Remove the first chanels
spectrum <- slice_signal(spectrum)

## Set peak positions (chanel) and expected energy values
calib_lines <- list(
  chanel = c(86, 496, 876),
  energy = c(238, 1461, 2614.5)
)

## Correct the energy scale
(spc1 <- calibrate_energy(spectrum, lines = calib_lines))

## Inspect results
plot(spc1, xaxis = "energy", yaxis = "count") +
  ggplot2::geom_vline(xintercept = c(238, 1461, 2614.5), linetype = 3)
}
\author{
N. Frerebeau
}
\concept{energy}
