% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calibrateDose.R
\docType{methods}
\name{calibrateDose}
\alias{calibrateDose}
\alias{calibrateDose-method}
\alias{calibrateDose,GammaSpectra,list-method}
\alias{calibrateDose,GammaSpectra-method}
\title{Calibration}
\usage{
calibrateDose(object, dose, ...)

\S4method{calibrateDose}{GammaSpectra,list}(object, dose, noise,
  laboratory = "LAB", ...)
}
\arguments{
\item{object}{An object of class \linkS4class{GammaSpectra}.}

\item{dose}{A \code{\link{list}} of length-two numeric vectors.}

\item{...}{Extra parameters passed to \code{\link{integrateSignal}}.}

\item{noise}{A \code{\link{list}} of two numeric values.}

\item{laboratory}{A \code{\link{character}} string giving the laboratory
name.}
}
\value{
An object of class \linkS4class{CalibrationCurve}.
}
\description{
Builds a calibration curve for gamma dose rate estimation.
}
\examples{
# Import CNF files for calibration
dir <- system.file("extdata/cerege/", package = "gamma")
calib_spc <- read(dir)[-5]

# Calibrate energy scale
calib_lines <- list(
  Pb = c(86, 238),
  K = c(493, 1461),
  Cs = c(876, 2614.5)
)

calib_scale <- calibrateEnergy(calib_spc, lines = calib_lines)

# Build calibration curve
known_dose <- list(
  BRIQUE = c(1984.64, 34.08),
  C347 = c(1421.38, 25.25),
  C341 = c(849.01, 21.32),
  PEP = c(2535.53, 112.17),
  GOU = c(1573.41, 17.43)
)

calib_curve <- calibrateDose(
  calib_spc,
  dose = known_dose,
  noise = list(value = 1190, error = 1)
)

# Plot curve
plot(calib_curve) +
  ggplot2::labs(x = "Dose rate [µGy/y]", y = "Integrated signal") +
  ggplot2::theme_bw()

# Import CNF file
dir <- system.file("extdata/", package = "gamma")
gamma_spc <- read(dir)

# Estimate gamma dose rate
dose_rate <- estimateDoseRate(gamma_spc, calib_curve)
}
\seealso{
\link{estimateDoseRate}
}
\author{
N. Frerebeau
}
