% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calibrateEnergy.R
\docType{methods}
\name{calibrateEnergy}
\alias{calibrateEnergy}
\alias{calibrateEnergy-method}
\alias{calibrateEnergy,GammaSpectra,data.frame-method}
\alias{calibrateEnergy,GammaSpectrum,data.frame-method}
\alias{calibrateEnergy,PeakModel,numeric-method}
\title{Spectrum calibration}
\usage{
calibrateEnergy(object, lines, ...)

\S4method{calibrateEnergy}{GammaSpectra,data.frame}(object, lines, ...)

\S4method{calibrateEnergy}{GammaSpectrum,data.frame}(object, lines, ...)

\S4method{calibrateEnergy}{PeakModel,numeric}(object, lines, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{GammaSpectrum}.}

\item{lines}{A \code{\link{numeric}} vector giving the peak positions
(in keV) of known sources. The default are 214Pb (238 keV), 40K (1461 keV)
and Cs (2614.5 keV).
Change it only if you know what you are doing.}

\item{...}{Extra parameters passed to \code{\link{removeBaseline}}.}
}
\value{
An object of class \linkS4class{GammaSpectrum}.
}
\description{
Calibrate the energy scale of a gamma spectrum.
}
\details{
TODO
}
\author{
N. Frerebeau
}
