% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CalibrationCurve-class}
\alias{CalibrationCurve-class}
\alias{DoseRateModel-class}
\alias{.DoseRateModel}
\alias{.CalibrationCurve}
\title{An S4 class to Represent a Dose Rate Calibration Curve}
\description{
An S4 class to Represent a Dose Rate Calibration Curve
}
\section{Slots}{

\describe{
\item{\code{model}}{A \code{\link[stats:lm]{linear model}} specifying the calibration
curve.}

\item{\code{background}}{A length-two \code{\link{numeric}} vector giving the
background noise value and error (see \code{\link{integrate_signal}}).}

\item{\code{range}}{A length-two \code{\link{numeric}} vector giving the energy
range to integrate within (see \code{\link{integrate_signal}}).}

\item{\code{Ni}}{A \linkS4class{DoseRateModel} object.}

\item{\code{NiEi}}{A \linkS4class{DoseRateModel} object.}

\item{\code{data}}{A \code{\link{data.frame}} giving the data used for
linear model fitting.}

\item{\code{details}}{A \code{\link{list}} of length-one vector giving the curve
metadata.}
}}

\section{Subset}{

In the code snippets below, \code{x} is a \code{CalibrationCurve} object.
\describe{
 \item{\code{x[[i]]}}{Extracts informations from a slot selected by
 subscript \code{i}. \code{i} is a \code{character} vector of length one.}
}
}

\seealso{
Other calibration class: 
\code{\link{PeakPosition-class}}
}
\author{
N. Frerebeau
}
\concept{calibration class}
