% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clermont}
\alias{clermont}
\title{Clermont Reference Data}
\format{TODO}
\source{
Miallier, D., Guérin, G., Mercier, N., Pilleyre, T. & Sanzelle, S.
 (2009). The Clermont Radiometric Reference Rocks: A Convenient Tool
 for Dosimetric Purposes. \emph{Ancient TL}, 27(2), p. 37-44.
}
\usage{
clermont
}
\description{
Clermont Reference Data
}
\keyword{datasets}
