% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{.decay}
\alias{.decay}
\title{Nuclear Decay Data}
\format{A \eqn{193 \times 13}{193 x 13} \code{\link[=data.frame]{data frame}}
 with the following columns (absolute errors):
 \describe{
   \item{decay_chain}{}
   \item{isotope}{}
   \item{occurrence}{}
   \item{occurrence_error}{}
   \item{post_radon}{}
   \item{half_life}{}
   \item{half_life_error}{}
   \item{energy}{Energy in keV.}
   \item{energy_error}{}
   \item{gamma_intensity}{}
   \item{gamma_intensity_error}{}
   \item{counts_chain}{}
   \item{counts_chain_error}{}
 }}
\source{
\href{IAEA Live Chart of Nuclides}{https://www-nds.iaea.org/relnsd/vcharthtml/VChartHTML.html}.
 Accessed 2019-06-07.
}
\usage{
.decay
}
\description{
An internal dataset containing the decay data for the following isotopes:
232-Th, 235-U, 238-U and 40-K.
}
\seealso{
Other datasets: \code{\link{.decay_La}},
  \code{\link{AIX100}}, \code{\link{BDX100}},
  \code{\link{BDX200}}, \code{\link{BDX300}},
  \code{\link{clermont}}
}
\concept{datasets}
\keyword{datasets}
\keyword{internal}
