% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/doserate.R
\docType{methods}
\name{doserate}
\alias{doserate}
\alias{fit_dose}
\alias{fit_dose-method}
\alias{predict_dose}
\alias{predict_dose-method}
\alias{fit_dose,GammaSpectra,GammaSpectrum-method}
\alias{fit_dose,GammaSpectra,numeric-method}
\alias{predict_dose,CalibrationCurve,missing-method}
\alias{predict_dose,CalibrationCurve,GammaSpectrum-method}
\alias{predict_dose,CalibrationCurve,GammaSpectra-method}
\title{Dose Rate Estimation}
\usage{
fit_dose(object, noise, ...)

predict_dose(object, spectrum, ...)

\S4method{fit_dose}{GammaSpectra,GammaSpectrum}(object, noise, range,
  intercept = TRUE, weights = FALSE, details = NULL, ...)

\S4method{fit_dose}{GammaSpectra,numeric}(object, noise, range,
  intercept = TRUE, weights = FALSE, details = NULL, ...)

\S4method{predict_dose}{CalibrationCurve,missing}(object, epsilon = 0,
  simplify = FALSE, ...)

\S4method{predict_dose}{CalibrationCurve,GammaSpectrum}(object, spectrum,
  epsilon = 0, simplify = FALSE, ...)

\S4method{predict_dose}{CalibrationCurve,GammaSpectra}(object, spectrum,
  epsilon = 0, simplify = FALSE, ...)
}
\arguments{
\item{object}{A \linkS4class{GammaSpectra} or \linkS4class{CalibrationCurve}
object.}

\item{noise}{A \linkS4class{GammaSpectrum} object or a length-two
\code{\link{numeric}} vector giving the background noise integration
value and error, respectively.}

\item{...}{Currently not used.}

\item{spectrum}{An optional \linkS4class{GammaSpectrum} or
\linkS4class{GammaSpectra} object in which to look for variables with which
to predict. If omitted, the fitted values are used.}

\item{range}{A length-two \code{\link{numeric}} vector giving the energy
range to integrate within (in keV).}

\item{intercept}{A \code{\link{logical}} scalar: should the intercept of the
curve be estimated? Change this only if you know what you are doing.}

\item{weights}{A \code{\link{logical}} scalar: should weights be used in the
fitting process. If \code{TRUE}, the inverse of the squared dose rate
errors are used. Change this only if you know what you are doing.}

\item{details}{A \code{\link{list}} of length-one vector specifying
additional informations about the instrument for which the curve is built.}

\item{epsilon}{A \code{\link{numeric}} value giving an extra error term
introduced by the calibration of the energy scale of the spectrum.}

\item{simplify}{A \code{\link{logical}} scalar: should the result be
simplified to a matrix? If \code{FALSE} (default), returns a list.}
}
\value{
A \linkS4class{CalibrationCurve} object.
}
\description{
\code{fit_dose} builds a calibration curve for gamma dose rate estimation.
}
\details{
\code{predict_dose} predicts in-situ gamma dose rate.
}
\examples{
# Import CNF files for calibration
spc_dir <- system.file("extdata/BDX100/calibration", package = "gamma")
spectra <- read(spc_dir)

# Set dose rate values and errors for each spectrum
data("clermont")
set_dose(spectra) <- clermont[, c("gamma", "gamma_error")]

# Build the calibration curve
calib_curve <- fit_dose(
  spectra,
  noise = c(25279.63171, 1.66235),
  range = c(165, 2800)
)

# Check the linear model
summary(calib_curve[["model"]])

# Plot the curve
plot(calib_curve) +
  ggplot2::labs(x = "Signal", y = "Dose rate [µGy/y]")

# Estimate gamma dose rates
(dose_rate <- predict_dose(calib_curve, spectra))
}
\references{
Mercier, N. & Falguères, C. (2007). Field Gamma Dose-Rate Measurement with
 a NaI(Tl) Detector: Re-Evaluation of the "Threshold" Technique.
 \emph{Ancient TL}, 25(1), p. 1-4.
}
\seealso{
\link{integrate_signal}
}
\author{
N. Frerebeau
}
\concept{dose rate}
