% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/extract.R
\docType{class}
\name{DoseRate-class}
\alias{DoseRate-class}
\alias{[[,DoseRate-method}
\title{An S4 class to represent a gamma dose rate (in µGy/y)}
\usage{
\S4method{[[}{DoseRate}(x, i)
}
\arguments{
\item{x}{An object of class \code{DoseRate}.}

\item{i}{A length-one \code{\link{character}} vector specifying the element
to extract or replace (see below). Character sring will be matched to the
names of the slots.}
}
\description{
An S4 class to represent a gamma dose rate (in µGy/y)
}
\section{Slots}{

\describe{
\item{\code{reference}}{A \code{\link{character}} vector giving the spectrum
references.}

\item{\code{dose_value}}{A \code{\link{numeric}} vector giving the dose rate
estimation.}

\item{\code{dose_error}}{A \code{\link{numeric}} vector giving the dose rate
estimation error.}

\item{\code{signal_value}}{A \code{\link{numeric}} vector giving the integrated
signal value.}

\item{\code{signal_error}}{A \code{\link{numeric}} vector giving the integrated
signal error.}
}}

\section{Coerce}{

In the code snippets below, \code{x} is a \code{DoseRate} object.
\describe{
 \item{\code{as(x, "matrix")}}{Coerces \code{x} to a \code{\link{matrix}}.}
 \item{\code{as(x, "data.frame")}}{Coerces \code{x} to a
 \code{\link[=data.frame]{data frame}}.}
}
}

\section{Subset}{

In the code snippets below, \code{x} is a \code{DoseRate} object.
\describe{
 \item{\code{x[[i]]}}{Extracts informations from a slot selected by
 subscript \code{i}. \code{i} is a \code{character} vector
 of length one.}
}
}

\author{
N. Frerebeau
}
