% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/energy.R
\docType{methods}
\name{energy}
\alias{energy}
\alias{calibrate_energy}
\alias{calibrate_energy-method}
\alias{is_calibrated}
\alias{calibrate_energy,GammaSpectrum,list-method}
\alias{calibrate_energy,GammaSpectrum,PeakPosition-method}
\alias{is_calibrated,GammaSpectrum-method}
\alias{is_calibrated,GammaSpectra-method}
\title{Energy Scale Calibration}
\usage{
calibrate_energy(object, lines, ...)

is_calibrated(object)

\S4method{calibrate_energy}{GammaSpectrum,list}(object, lines, ...)

\S4method{calibrate_energy}{GammaSpectrum,PeakPosition}(object, lines, ...)

\S4method{is_calibrated}{GammaSpectrum}(object)

\S4method{is_calibrated}{GammaSpectra}(object)
}
\arguments{
\item{object}{A \linkS4class{GammaSpectrum} or
\linkS4class{GammaSpectra} object.}

\item{lines}{A \linkS4class{PeakPosition} object or a \code{\link{list}} of
length two. If a \code{list} is provided, each element must be a named
numeric vector giving the observed peak position ("\code{chanel}") and the
corresponding expected "\code{energy}" value (in keV).}

\item{...}{Currently not used.}
}
\value{
\code{calibrate_energy} returns a \linkS4class{GammaSpectrum} object.

 \code{is_calibrated} returns a \code{\link{logical}} vector.
}
\description{
Calibrates the energy scale of a gamma spectrum.
}
\examples{
## Import a CNF file for calibration
spc_file <- system.file("extdata/test_LaBr.TKA", package = "gamma")
(spectrum <- read(spc_file))

## Set peak positions (chanel) and expected energy values
calib_lines <- list(
  chanel = c(86, 495, 879),
  energy = c(238, 1461, 2615)
)

## Correct the energy scale
(spc1 <- calibrate_energy(spectrum, lines = calib_lines))

## Inspect results
plot(spc1, xaxis = "energy", yaxis = "count") +
  ggplot2::geom_vline(xintercept = c(238, 1461, 2614.5), linetype = 3)
}
\author{
N. Frerebeau
}
\concept{energy}
