% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/energy_calibrate.R
\docType{methods}
\name{energy}
\alias{energy}
\alias{energy_calibrate}
\alias{energy_calibrate-method}
\alias{has_energy}
\alias{energy_calibrate,GammaSpectrum,list-method}
\alias{energy_calibrate,GammaSpectrum,PeakPosition-method}
\alias{has_energy,GammaSpectrum-method}
\alias{has_energy,GammaSpectra-method}
\title{Energy Scale Calibration}
\usage{
energy_calibrate(object, lines, ...)

has_energy(object)

\S4method{energy_calibrate}{GammaSpectrum,list}(object, lines, ...)

\S4method{energy_calibrate}{GammaSpectrum,PeakPosition}(object, lines, ...)

\S4method{has_energy}{GammaSpectrum}(object)

\S4method{has_energy}{GammaSpectra}(object)
}
\arguments{
\item{object}{A \linkS4class{GammaSpectrum} or
\linkS4class{GammaSpectra} object.}

\item{lines}{A \linkS4class{PeakPosition} object or a \code{\link{list}} of
length two. If a \code{list} is provided, each element must be a named
numeric vector giving the observed peak position ("\code{channel}") and the
corresponding expected "\code{energy}" value (in keV).}

\item{...}{Currently not used.}
}
\value{
\code{energy_calibrate} returns a \linkS4class{GammaSpectrum} object.

 \code{has_energy} returns a \code{\link{logical}} vector.
}
\description{
Calibrates the energy scale of a gamma spectrum.
}
\examples{
## Import a CNF file
spc_file <- system.file("extdata/LaBr.TKA", package = "gamma")
(spc <- read(spc_file))

## Set peak positions (channel) and expected energy values
calib_lines <- list(
  channel = c(86, 495, 879),
  energy = c(238, 1461, 2615)
)

## Adjust the energy scale
(spc1 <- energy_calibrate(spc, lines = calib_lines))

## Inspect results
plot(spc1, xaxis = "energy", yaxis = "count") +
  ggplot2::geom_vline(xintercept = c(238, 1461, 2614.5), linetype = 3)
}
\author{
N. Frerebeau
}
\concept{energy}
