% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/estimateDoseRate.R
\docType{methods}
\name{estimateDoseRate}
\alias{estimateDoseRate}
\alias{estimateDoseRate-method}
\alias{estimateDoseRate,GammaSpectrum,CalibrationCurve-method}
\alias{estimateDoseRate,GammaSpectrum-method}
\alias{estimateDoseRate,GammaSpectra,CalibrationCurve-method}
\alias{estimateDoseRate,GammaSpectra-method}
\title{Gamma dose rate}
\usage{
estimateDoseRate(object, curve, ...)

\S4method{estimateDoseRate}{GammaSpectrum,CalibrationCurve}(object, curve,
  noise, ...)

\S4method{estimateDoseRate}{GammaSpectra,CalibrationCurve}(object, curve,
  noise, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{GammaSpectra}.}

\item{curve}{An object of class \linkS4class{CalibrationCurve}.}

\item{...}{Extra parameters passed to \code{\link{integrateSignal}}.}

\item{noise}{A \code{\link{list}} of numeric values.}
}
\description{
Estimates in-situ gamma dose rate.
}
\examples{
# Import CNF files for calibration
dir <- system.file("extdata/cerege/", package = "gamma")
calib_spc <- read(dir)[-5]

# Build calibration curve
known_dose <- list(
  BRIQUE = c(1984.64, 34.08),
  C347 = c(1421.38, 25.25),
  C341 = c(849.01, 21.32),
  PEP = c(2535.53, 112.17),
  GOU = c(1573.41, 17.43)
)

calib_curve <- calibrate(
  calib_spc,
  dose = known_dose,
  noise = list(value = 1190, error = 1),
  span = 20
)

# Plot curve
plot(calib_curve) +
  ggplot2::labs(x = "Dose rate [µGy/y]", y = "Integrated signal") +
  ggplot2::theme_bw()

# Import CNF file
dir <- system.file("extdata/", package = "gamma")
gamma_spc <- read(dir)

# Estimate gamma dose rate
dose_rate <- estimateDoseRate(gamma_spc, calib_curve,
                              noise = list(value = 1190, error = 1))

# Plot estimated values on the calibration curve
plot(calib_curve, dose_rate)
}
\author{
N. Frerebeau
}
