% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/findPeaks.R
\docType{methods}
\name{findPeaks}
\alias{findPeaks}
\alias{findPeaks-method}
\alias{findPeaks,GammaSpectra-method}
\alias{findPeaks,GammaSpectrum-method}
\title{Find peaks}
\usage{
findPeaks(object, ...)

\S4method{findPeaks}{GammaSpectra}(object, span = NULL, ...)

\S4method{findPeaks}{GammaSpectrum}(object, span = NULL, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{GammaSpectrum}.}

\item{...}{Currently not used.}

\item{span}{A \code{\link{numeric}} giving the half window size. If
\code{NULL}, 5\% of the number of chanels is used as the half window size.}
}
\value{
A three columns data frame or a list of data frame.
}
\description{
Finds local maxima in sequential data.
}
\details{
A local maximum have to be the highest one in the given window to be
 recognized as peak.
}
\note{
Adapted from Stasia Grinberg's \href{https://github.com/stas-g/findPeaks}{algorithm}.
}
\examples{
# Import a CNF file
spc_dir <- system.file("extdata/test1.cnf", package = "gamma")
spc_dir <- system.file("extdata/cerege", package = "gamma")
spc_test <- read(spc_dir)[[5]]

# Estimate baseline
baseline <- estimateBaseline(spc_test)

# Plot spectrum + baseline
plot(spc_test) +
  ggplot2::geom_line(colour = "red", data = baseline,
                     mapping = ggplot2::aes(x = energy, y = baseline))

# Remove baseline
spc_clean <- removeBaseline(spc_test)

# Detect peaks
peak_index <- findPeaks(spc_clean, span = NULL)

plot(spc_clean, select = 1:3, facet = TRUE) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts") +
  ggplot2::geom_vline(xintercept = peak_index$energy, colour = "red") +
  ggplot2::theme_bw()
}
\author{
N. Frerebeau
}
