% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/doserate.R
\docType{methods}
\name{fit_dose}
\alias{fit_dose}
\alias{fit_dose-method}
\alias{fit_dose,GammaSpectra,GammaSpectrum-method}
\alias{fit_dose,GammaSpectra,numeric-method}
\title{Dose Rate Estimation}
\usage{
fit_dose(object, noise, ...)

\S4method{fit_dose}{GammaSpectra,GammaSpectrum}(object, noise, range,
  intercept = TRUE, weights = FALSE, details = NULL, ...)

\S4method{fit_dose}{GammaSpectra,numeric}(object, noise, range,
  intercept = TRUE, weights = FALSE, details = NULL, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{GammaSpectra}.}

\item{noise}{A \code{\link{list}} of two numeric values giving the noise
value and error.}

\item{...}{Currently not used.}

\item{range}{A length-two \code{\link{numeric}} vector giving the energy
range to integrate within (in keV).}

\item{intercept}{A \code{\link{logical}} scalar: should the intercept of the
curve be estimated?}

\item{weights}{A \code{\link{logical}} scalar: should weights be used in the
fitting process. If \code{TRUE}, the inverse of the squared dose errors are
used.}

\item{details}{A list of \code{\link{character}} vector specifying additional
informations about the instrument for which the curve is built.}
}
\value{
A \linkS4class{CalibrationCurve} object.
}
\description{
Builds a calibration curve for gamma dose rate estimation.
}
\details{
TODO
}
\examples{
# Import CNF files for calibration
spc_dir <- system.file("extdata/BDX100/calibration", package = "gamma")
spectra <- read(spc_dir)

# Set dose rate values and errors for each spectrum
data("clermont")
set_dose(spectra) <- clermont[, c("gamma", "gamma_error")]

# Build the calibration curve
calib_curve <- fit_dose(
  spectra,
  noise = c(25279.63171, 1.66235),
  range = c(165, 2800)
)

# Check the linear model
summary(calib_curve[["model"]])

# Plot the curve
plot(calib_curve) +
  ggplot2::labs(x = "Signal", y = "Dose rate [µGy/y]")

# Estimate gamma dose rates
(dose_rate <- predict_dose(calib_curve, spectra))
}
\seealso{
\link{integrate_signal}

Other dose rate: \code{\link{predict_dose}}
}
\author{
N. Frerebeau
}
\concept{dose rate}
