% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/fit.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit-method}
\alias{fit,GammaSpectra,GammaSpectrum-method}
\alias{fit,GammaSpectra,numeric-method}
\title{Calibration}
\usage{
fit(object, noise, ...)

\S4method{fit}{GammaSpectra,GammaSpectrum}(object, noise, range = c(200,
  2800), intercept = TRUE, weights = FALSE, details = NULL, ...)

\S4method{fit}{GammaSpectra,numeric}(object, noise, range = c(200, 2800),
  intercept = TRUE, weights = FALSE, details = NULL, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{GammaSpectra}.}

\item{noise}{A \code{\link{list}} of two numeric values giving the noise
value and error.}

\item{...}{Currently not used.}

\item{range}{A length-two \code{\link{numeric}} vector giving the energy
range to integrate within (in keV).}

\item{intercept}{A \code{\link{logical}} scalar: should the intercept of the
curve be estimated?}

\item{weights}{A \code{\link{logical}} scalar: should weights be used in the
fitting process. If \code{TRUE}, the inverse of the squared dose errors are
used.}

\item{details}{A list of \code{\link{character}} vector specifying additional
informations about the instrument for which the curve is built.}
}
\value{
A \linkS4class{CalibrationCurve} object.
}
\description{
Builds a calibration curve for gamma dose rate estimation.
}
\details{
TODO
}
\examples{
# Import CNF files for calibration
spc_dir <- system.file("extdata/crp2a/calibration", package = "gamma")
spc_calib <- read(spc_dir, skip = TRUE)

# Set dose rate values and errors for each spectrum
setDoseRate(spc_calib) <- list(
  BRIQUE = c(1986, 36),
  C341 = c(850, 21),
  C347 = c(1424, 24),
  GOU = c(1575, 17),
  LMP = c(642, 18),
  MAZ = c(1141, 12),
  PEP = c(2538, 112)
)

# Build the calibration curve
calib_curve <- fit(
  spc_calib,
  noise = c(value = 1190, error = 1),
  range = c(200, 2800)
)

# Check the linear model
summary(calib_curve[["model"]])

# Plot the curve
plot(calib_curve) +
  ggplot2::labs(x = "Signal", y = "Dose rate [µGy/y]")

# Estimate gamma dose rates
(dose_rate <- predict(calib_curve, spc_calib))
}
\seealso{
\link{integrateSignal}

Other dose rate: \code{\link{predict}}
}
\author{
N. Frerebeau
}
\concept{dose rate}
