% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma-package.R
\docType{package}
\name{gamma-package}
\alias{gamma-package}
\alias{_PACKAGE}
\alias{gamma}
\title{gamma: Dose Rate Estimation from in-Situ Gamma-Ray Spectrometry
    Measurements}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Process in-situ Gamma-Ray Spectrometry for
    Luminescence Dating. This package allows to import, inspect and
    correct the energy shifts of Gamma-ray spectra. It provides methods
    for estimating the gamma dose rate by the use of a calibration curve.
    The package only supports Canberra CNF and TKA files.
}
\details{
\tabular{ll}{
 \strong{Package:} \tab gamma \cr
 \strong{Type:} \tab Package \cr
 \strong{Version:} \tab 1.0.1 \cr
 \strong{License:} \tab GPL-3 \cr
 \strong{DOI:} \tab \href{https://doi.org/10.5281/zenodo.2652393}{10.5281/zenodo.2652393}
}
}
\note{
This work received a state financial support managed by the Agence Nationale
de la Recherche (France) through the program \emph{Investissements d'avenir}
(ref. \href{https://lascarbx.labex.u-bordeaux.fr}{ANR-10-LABX-52} and
\href{https://www.univ-amu.fr/amidex/}{11-IDEX-0001}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://crp2a.github.io/gamma/}
  \item \url{https://github.com/crp2a/gamma}
  \item Report bugs at \url{https://github.com/crp2a/gamma/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
 Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
 Christelle Lahaye \tab \emph{Université Bordeaux Montaigne, France} \cr
 Brice Lebrun \tab \emph{Université Bordeaux Montaigne, France} \cr
 Guilhem Paradol \tab \emph{Commissariat à l'Énergie Atomique, France} \cr
 Magali Rizza \tab \emph{Aix-Marseille Université, France}
}

\strong{Package maintainer}

Brice Lebrun\cr
\email{brice.lebrun@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
