% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GammaSpectrum-class}
\alias{GammaSpectrum-class}
\alias{.GammaSpectrum}
\title{An S4 Class to Represent a Gamma Sectrum}
\description{
Represents a single spectrum of a gamma ray spectrometry measurement.
}
\section{Slots}{

\describe{
\item{\code{hash}}{A \code{\link{character}} string giving the 32-byte MD5 hash of
the imported file.}

\item{\code{name}}{A \code{\link{character}} string the measurement reference.}

\item{\code{date}}{A \code{\link{POSIXct}} element giving the measurement date and
time.}

\item{\code{instrument}}{A \code{\link{character}} string giving the instrument name.}

\item{\code{file_format}}{A \code{\link{character}} string giving the format of the
imported file.}

\item{\code{live_time}}{A \code{\link{numeric}} value.}

\item{\code{real_time}}{A \code{\link{numeric}} value.}

\item{\code{chanel}}{A \code{\link{integer}} vector giving the channel number.
Numeric values are coerced to integer as by \code{\link{as.integer}}
(and hence truncated towards zero).}

\item{\code{energy}}{A \code{\link{numeric}} vector giving the gamma ray's energy
(in keV).}

\item{\code{count}}{A \code{\link{numeric}} vector giving the counts number for
each channel. Numeric values are coerced to integer as by
\code{\link{as.integer}} (and hence truncated towards zero).}

\item{\code{rate}}{A \code{\link{numeric}} vector the count rate (in 1/s) for
each channel.}

\item{\code{calibration}}{A \code{\link[stats:lm]{linear model}} used for energy
scale calibration (see \code{\link{calibrate_energy}}).}

\item{\code{dose_rate}}{A length-two \code{\link{numeric}} vector giving the dose
rate and corresponding error.}
}}

\section{Access}{

In the code snippets below, \code{x} is a \code{GammaSpectrum} object.
\describe{
 \item{\code{get_hash(x)}}{Get the MD5 hash of the raw data file.}
 \item{\code{get_chanels(x)}}{Get the number of chanels in \code{x}.}
 \item{\code{get_energy(x)}}{Get the energy range of \code{x}.}
 \item{\code{get_dose(x)}}{Get the dose rate of \code{x}.}
}
}

\section{Coerce}{

In the code snippets below, \code{x} is a \code{GammaSpectrum} object.
\describe{
 \item{\code{as(x, "matrix")}}{Coerces \code{x} to a \code{\link{matrix}}.}
 \item{\code{as(x, "data.frame")}}{Coerces \code{x} to a
 \code{\link[=data.frame]{data frame}}.}
}
}

\section{Subset}{

In the code snippets below, \code{x} is a \code{GammaSpectrum} object.
\describe{
 \item{\code{x[[i]]}}{Extracts informations from a slot selected by
 subscript \code{i}. \code{i} is a \code{character} vector
 of length one and will be matched to the name of the slots.}
}
}

\examples{
# Import a Canberra CNF file
spc_file <- system.file("extdata/test_CNF.cnf", package = "gamma")
(spectrum <- read(spc_file))

## Access
get_hash(spectrum)
get_chanels(spectrum)
get_energy(spectrum)
get_dose(spectrum)

# Coerce
df <- as(spectrum, "data.frame")
head(df)
tail(df)

# Subset
spectrum[["hash"]]
spectrum[["name"]]
spectrum[["date"]]
spectrum[["instrument"]]
spectrum[["file_format"]]
spectrum[["live_time"]]
spectrum[["real_time"]]
spectrum[["calibration"]]
}
\seealso{
\linkS4class{GammaSpectra}, \linkS4class{BaseLine}

Other class: \code{\link{BaseLine-class}},
  \code{\link{CalibrationCurve-class}},
  \code{\link{GammaSpectra-class}},
  \code{\link{PeakPosition-class}}
}
\author{
N. Frerebeau
}
\concept{class}
