% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heidelberg_2024}
\alias{heidelberg_2024}
\title{Heidelberg Reference Data 2024}
\format{
A \code{\link{data.frame}} with tabulated data and conversion factor reference.
}
\source{
Adamiec, G. & Aitken, M.J. (1998).
Dose-rate conversion factors: update. \emph{Ancient TL}, 16, p. 37–50.

Cresswell, A.J., Carter, J. & Sanderson, D.C.W. (2018). Dose rate
conversion parameters: Assessment of nuclear data.
\emph{Radiation Measurements}, 120, p. 195–201.

Guérin, G., Mercier, N. & Adamiec, G. (2011). Dose-Rate Conversion Factors:
Update. \emph{Ancient TL}, 29(1), p. 5-8.

Liritzis, I., Stamoulis, K., Papachristodoulou, C. & Ioannides, K. (2013).
A re-evaluation of radiation dose-rate conversion factors.
\emph{Mediterranean Archaeology and Archaeometry}, 12, p. 1–15.

Miallier, D., Guérin, G., Mercier, N., Pilleyre, T. & Sanzelle, S.
(2009). The Clermont Radiometric Reference Rocks: A Convenient Tool
for Dosimetric Purposes. \emph{Ancient TL}, 27(2), p. 37-44.
}
\usage{
data("heidelberg_2024")
}
\description{
Reference data for the calibration sites near Heidelberg as detailed
in Kreutzer at al. (2025)
}
\references{
Kreutzer, S., Martin, L., Miallier, D., Mercier, N., 2025.
Environmental gamma dose rate measurements using cadmium zinc telluride (CZT) detectors.
Geochronology 7, 229–246. \doi{10.5194/gchron-7-229-2025}
}
\seealso{
Other datasets: 
\code{\link{AIX_NaI_1}},
\code{\link{BDX_LaBr_1}},
\code{\link{clermont}},
\code{\link{clermont_2024}}
}
\concept{datasets}
\keyword{datasets}
