% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/integrateSignal.R
\docType{methods}
\name{integrateSignal}
\alias{integrateSignal}
\alias{integrateSignal-method}
\alias{integrateSignal,GammaSpectra-method}
\alias{integrateSignal,GammaSpectrum-method}
\title{Signal integration}
\usage{
integrateSignal(object, ...)

\S4method{integrateSignal}{GammaSpectra}(object, range = c(200, 2800),
  noise = NULL, NiEi = TRUE, ..., simplify = FALSE)

\S4method{integrateSignal}{GammaSpectrum}(object, range = c(200, 2800),
  noise = NULL, NiEi = TRUE, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{GammaSpectrum}.}

\item{...}{Currently not used.}

\item{range}{A length two \code{\link{numeric}} vector giving the energy
range to integrate within (in keV).}

\item{noise}{A length-two \code{\link{numeric}} vector.}

\item{NiEi}{A \code{\link{logical}} scalar.}

\item{simplify}{A \code{\link{logical}} scalar: should the result be
simplified to a matrix? The default value, \code{FALSE}, returns a list.}
}
\value{
If \code{simplify} is \code{FALSE}, then returns a list length-two
 numeric vectors (default), else returns a matrix.
}
\description{
Signal integration
}
\details{
TODO
}
\author{
N. Frerebeau
}
\keyword{internal}
