% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calibrate.R
\docType{methods}
\name{integrateSignal}
\alias{integrateSignal}
\alias{integrateSignal-method}
\alias{integrateSignal,GammaSpectra-method}
\alias{integrateSignal,GammaSpectrum-method}
\title{Signal integration}
\usage{
integrateSignal(object, ...)

\S4method{integrateSignal}{GammaSpectra}(object, peaks = c(238, 1461,
  2614.5), noise = list(value = 1190, error = 1), m = 3, ...)

\S4method{integrateSignal}{GammaSpectrum}(object, peaks = c(238, 1461,
  2614.5), noise = list(value = 1190, error = 1), m = 3, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{GammaSpectrum}.}

\item{...}{Currently not used.}

\item{peaks}{A \code{\link{numeric}} vector.}

\item{noise}{A \code{\link{list}} of numeric values.}

\item{m}{An \code{\link{integer}}.}
}
\description{
Signal integration
}
\author{
N. Frerebeau
}
\keyword{internal}
