% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/integrateSignal.R
\docType{methods}
\name{integrateSignal}
\alias{integrateSignal}
\alias{integrateSignal-method}
\alias{integrateSignal,GammaSpectra,numeric,numeric-method}
\alias{integrateSignal,GammaSpectra,numeric,missing-method}
\alias{integrateSignal,GammaSpectra,numeric,GammaSpectrum-method}
\alias{integrateSignal,GammaSpectrum,numeric,missing-method}
\alias{integrateSignal,GammaSpectrum,numeric,numeric-method}
\alias{integrateSignal,GammaSpectrum,numeric,GammaSpectrum-method}
\title{Signal integration}
\usage{
integrateSignal(object, range, noise, ...)

\S4method{integrateSignal}{GammaSpectra,numeric,numeric}(object, range,
  noise, NiEi = TRUE, simplify = FALSE, ...)

\S4method{integrateSignal}{GammaSpectra,numeric,missing}(object, range,
  NiEi = TRUE, simplify = FALSE, ...)

\S4method{integrateSignal}{GammaSpectra,numeric,GammaSpectrum}(object,
  range, noise, NiEi = TRUE, simplify = FALSE, ...)

\S4method{integrateSignal}{GammaSpectrum,numeric,missing}(object, range,
  NiEi = TRUE, ...)

\S4method{integrateSignal}{GammaSpectrum,numeric,numeric}(object, range,
  noise, NiEi = TRUE, ...)

\S4method{integrateSignal}{GammaSpectrum,numeric,GammaSpectrum}(object,
  range, noise, NiEi = TRUE, ...)
}
\arguments{
\item{object}{A \linkS4class{GammaSpectrum} or \linkS4class{GammaSpectra}
object.}

\item{range}{A length-two \code{\link{numeric}} vector giving the energy
range to integrate within (in keV).}

\item{noise}{A length-two \code{\link{numeric}} vector giving the noise
value and error, respectively (see details).}

\item{...}{Currently not used.}

\item{NiEi}{A \code{\link{logical}} scalar.
Change this only if you know what you are doing.}

\item{simplify}{A \code{\link{logical}} scalar: should the result be
simplified to a matrix? The default value, \code{FALSE}, returns a list.}
}
\value{
If \code{simplify} is \code{FALSE} returns a list of length-two
 numeric vectors (default), else returns a matrix.
}
\description{
Signal integration
}
\details{
TODO

 It assumes that each spectrum is calibrated in energy.
}
\references{
Guérin, G. & Mercier, M. (2011). Determining Gamma Dose Rates by Field Gamma
 Spectroscopy in Sedimentary Media: Results of Monte Carlo Simulations.
 \emph{Radiation Measurements}, 46(2), p. 190-195.
 DOI: \href{https://www.doi.org/10.1016/j.radmeas.2010.10.003}{10.1016/j.radmeas.2010.10.003}.

 Mercier, N. & Falguères, C. (2007). Field Gamma Dose-Rate Measurement with
 a NaI(Tl) Detector: Re-Evaluation of the "Threshold" Technique.
 \emph{Ancient TL}, 25(1), p. 1-4.
}
\seealso{
Other signal processing: \code{\link{peaks}},
  \code{\link{processBaseline}}, \code{\link{smooth}},
  \code{\link{stabilize}}
}
\author{
N. Frerebeau
}
\concept{signal processing}
\keyword{internal}
