% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Shiny App}
\usage{
launch_app()
}
\value{
A \pkg{shiny} application object.
}
\description{
A wrapper for \code{\link[shiny]{shinyApp}}.
}
\examples{
\dontrun{launch_app()}
}
\seealso{
Other IO: \code{\link{operators}}, \code{\link{plot}},
  \code{\link{read}}, \code{\link{summarise}}
}
\author{
N. Frerebeau
}
\concept{IO}
