% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/extract.R
\docType{class}
\name{PeakModel-class}
\alias{PeakModel-class}
\alias{[[,PeakModel-method}
\title{An S4 class to represent a set peaks}
\usage{
\S4method{[[}{PeakModel}(x, i)
}
\arguments{
\item{x}{An object of class \code{PeakModel}.}

\item{i}{A length-one \code{\link{character}} vector specifying the element
to extract or replace (see below). Character sring will be matched to the
names of the slots.}
}
\description{
An S4 class to represent a set peaks
}
\section{Slots}{

\describe{
\item{\code{model}}{A list of \code{\link[stats:nls]{nonlinear models}} giving the
fitted model for each peak.}

\item{\code{coefficients}}{A \code{\link{numeric}} matrix giving the peak parameters.}

\item{\code{spectrum}}{A \linkS4class{GammaSpectrum} object.}

\item{\code{baseline}}{A \linkS4class{BaseLine} object.}
}}

\section{Subset}{

In the code snippets below, \code{x} is a \code{PeakModel} object.
\describe{
 \item{\code{x[[i]]}}{Extracts informations from a slot selected by
 subscript \code{i}. \code{i} is a \code{character} vector
 of length one.}
}
}

\author{
N. Frerebeau
}
