% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PeakPosition-class}
\alias{PeakPosition-class}
\alias{.PeakPosition}
\title{An S4 Class to Represent a Set of Peaks}
\description{
An S4 Class to Represent a Set of Peaks
}
\section{Slots}{

\describe{
\item{\code{hash}}{TODO.}

\item{\code{noise_method}}{A \code{\link{character}} string specifying the method
used for peak detection.}

\item{\code{noise_threshold}}{A length one \code{\link{numeric}} vector giving the
noise threshold.}

\item{\code{window}}{A length one \code{\link{numeric}} vector giving the half-window
size.}

\item{\code{chanel}}{TODO.}

\item{\code{energy}}{TODO.}
}}

\section{Subset}{

In the code snippets below, \code{x} is a \code{PeakPosition} object.
\describe{
 \item{\code{x[[i]]}}{Extracts informations from a slot selected by
 subscript \code{i}. \code{i} is a \code{character} vector
 of length one.}
}
}

\seealso{
Other class: \code{\link{BaseLine-class}},
  \code{\link{CalibrationCurve-class}},
  \code{\link{GammaSpectra-class}},
  \code{\link{GammaSpectrum-class}}
}
\author{
N. Frerebeau
}
\concept{class}
