% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/peaks.R
\docType{methods}
\name{peaks}
\alias{peaks}
\alias{findPeaks}
\alias{findPeaks-method}
\alias{fitPeaks}
\alias{fitPeaks-method}
\alias{findPeaks,GammaSpectrum-method}
\alias{fitPeaks,GammaSpectrum,numeric-method}
\alias{fitPeaks,PeakPosition,missing-method}
\title{Peaks}
\usage{
findPeaks(object, ...)

fitPeaks(object, peaks, ...)

\S4method{findPeaks}{GammaSpectrum}(object, method = c("MAD"), SNR = 2,
  span = NULL, ...)

\S4method{fitPeaks}{GammaSpectrum,numeric}(object, peaks,
  scale = c("chanel", "energy"), bounds = NULL, ...)

\S4method{fitPeaks}{PeakPosition,missing}(object, scale = c("energy",
  "chanel"), bounds = NULL, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{GammaSpectrum} or
\linkS4class{PeakPosition}.}

\item{...}{Extra parameters passed to \code{\link{estimateBaseline}}.}

\item{peaks}{A \code{\link{numeric}} vector giving the starting peak postions
for the nonlinear model fitting (see below).}

\item{method}{A \code{\link{character}} string specifying the methode to be
used for background noise estimation (see below).}

\item{SNR}{An \code{\link{integer}} giving the signal-to-noise-ratio for
peak detection (see below).}

\item{span}{An \code{\link{integer}} giving the half window size (in number
of chanels). If \code{NULL}, 5\% of the number of chanels is used as the
half window size.}

\item{scale}{A \code{\link{character}} string specifying the scale of
\code{peaks}. It must be one of "\code{chanel}" (the default) or
"\code{energy}".}

\item{bounds}{A \code{\link{numeric}} vector giving the
parameters bounds (in percent), replicated to be of length three (see
below). If \code{NULL} (the default), all parameters are assumed to be
unconstrained.}
}
\value{
\code{findPeaks} returns an object of class \linkS4class{PeakPosition}.

 \code{fitPeaks} returns an object of class \linkS4class{PeakModel}.
}
\description{
Finds local maxima in sequential data.
}
\section{Peak detection}{

 A local maximum has to be the highest one in the given window and has to be
 higher than \eqn{SNR \times noise}{SNR * noise} to be recognized as peak.
}

\section{Peak fitting}{

 TODO
}

\examples{
# Import a Canberra CNF file
spc_file <- system.file("extdata/test.cnf", package = "gamma")
spectrum <- read(spc_file, skip = TRUE)

# Find peaks
(peaks <- findPeaks(spectrum))
plot(peaks)

# Fit peaks
## Using previously detected peaks
(fit1 <- fitPeaks(peaks))
plot(fit1)
## Specifying starting positions
(fit2 <- fitPeaks(spectrum, peaks = c(86, 496, 876), scale = "chanel"))
plot(fit2)
}
\seealso{
\link{estimateBaseline}
}
\author{
N. Frerebeau
}
