% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/peaks.R
\docType{methods}
\name{find_peaks}
\alias{find_peaks}
\alias{find_peaks-method}
\alias{find_peaks,GammaSpectrum-method}
\title{Peaks}
\usage{
find_peaks(object, ...)

\S4method{find_peaks}{GammaSpectrum}(object, method = c("MAD"),
  SNR = 2, span = NULL, ...)
}
\arguments{
\item{object}{A \linkS4class{GammaSpectrum} or \linkS4class{PeakPosition}
object.}

\item{...}{Extra parameters to be passed to internal methods.}

\item{method}{A \code{\link{character}} string specifying the method to be
used for background noise estimation (see below).}

\item{SNR}{An \code{\link{integer}} giving the signal-to-noise-ratio for
peak detection (see below).}

\item{span}{An \code{\link{integer}} giving the half window size (in number
of chanels). If \code{NULL}, 5\% of the number of chanels is used as the
half window size.}
}
\value{
An object of class \linkS4class{PeakPosition}.
}
\description{
Finds local maxima in sequential data.
}
\details{
A local maximum has to be the highest one in the given window and has to be
 higher than \eqn{SNR \times noise}{SNR * noise} to be recognized as peak.

 The following methods are availble for noise estimation:
 \describe{
  \item{MAD}{Median Absolute Deviation.}
 }
}
\examples{
# Import a Canberra CNF file
spc_file <- system.file("extdata/test_LaBr.CNF", package = "gamma")
spectrum <- read(spc_file)

# Find peaks
(peaks <- find_peaks(spectrum))
plot(spectrum, peaks)
}
\seealso{
Other signal processing: \code{\link{baseline}},
  \code{\link{integrate_signal}},
  \code{\link{slice_signal}}, \code{\link{smooth_signal}},
  \code{\link{stabilize_signal}}
}
\author{
N. Frerebeau
}
\concept{signal processing}
