% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot-method}
\alias{plot,GammaSpectrum,missing-method}
\alias{plot,GammaSpectrum,GammaSpectrum-method}
\alias{plot,GammaSpectra,missing-method}
\alias{plot,GammaSpectrum,PeakPosition-method}
\alias{plot,CalibrationCurve,missing-method}
\title{Plot}
\usage{
\S4method{plot}{GammaSpectrum,missing}(x, y, ...)

\S4method{plot}{GammaSpectrum,GammaSpectrum}(x, y, xaxis = c("chanel",
  "energy"), yaxis = c("counts", "rate"), ...)

\S4method{plot}{GammaSpectra,missing}(x, xaxis = c("chanel", "energy"),
  yaxis = c("counts", "rate"), select = NULL, facet = FALSE, ...)

\S4method{plot}{GammaSpectrum,PeakPosition}(x, y, ...)

\S4method{plot}{CalibrationCurve,missing}(x, y, ...)
}
\arguments{
\item{x, y}{Objects to be plotted.}

\item{...}{Currently not used.}

\item{xaxis, yaxis}{A \code{\link{character}} string specifying the data to
be plotted along each axis. It must be one of "\code{energy}" or
"\code{chanel}" (\code{x} axis) and "\code{counts}" or "\code{rate}"
(\code{y} axis). Any unambiguous substring can be given.}

\item{select}{A \code{\link{numeric}} or \code{\link{character}} vector
giving the selection of the spectrum that are drawn.}

\item{facet}{A \code{\link{logical}} scalar: should a matrix of panels
defined by spectrum be drawn?}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot
}
\examples{
# Import CNF files
spc_dir <- system.file("extdata/crp2a/calibration", package = "gamma")
spectra <- read(spc_dir)

# Plot all spectra
plot(spectra, yaxis = "rate", facet = FALSE) +
  ggplot2::labs(x = "Energy [keV]", y = "Count rate [1/s]") +
  ggplot2::theme_bw()

# Plot the spectrum named 'BRIQUE'
plot(spectra, xaxis = "energy", yaxis = "count",
     select = "BRIQUE", facet = TRUE) +
  ggplot2::labs(x = "Energy [keV]", y = "Count") +
  ggplot2::theme_bw()

# Plot the first three spectra
plot(spectra, xaxis = "chanel", yaxis = "rate",
     select = 1:3, facet = TRUE) +
  ggplot2::labs(x = "Chanel", y = "Count rate [1/s]") +
  ggplot2::theme_bw()
}
\author{
N. Frerebeau
}
