% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot-method}
\alias{plot,GammaSpectrum,missing-method}
\alias{plot,GammaSpectrum,GammaSpectrum-method}
\alias{plot,GammaSpectrum,BaseLine-method}
\alias{plot,GammaSpectra,missing-method}
\alias{plot,PeakPosition,missing-method}
\alias{plot,PeakModel,missing-method}
\alias{plot,CalibrationCurve,missing-method}
\alias{plot,CalibrationCurve,DoseRate-method}
\title{Plot}
\usage{
\S4method{plot}{GammaSpectrum,missing}(x, xaxis = c("energy", "chanel"),
  yaxis = c("counts", "rate"), ...)

\S4method{plot}{GammaSpectrum,GammaSpectrum}(x, y, xaxis = c("energy",
  "chanel"), yaxis = c("counts", "rate"), ...)

\S4method{plot}{GammaSpectra,missing}(x, xaxis = c("energy", "chanel"),
  yaxis = c("counts", "rate"), select = NULL, facet = FALSE, ...)

\S4method{plot}{PeakPosition,missing}(x, xaxis = c("energy", "chanel"),
  yaxis = c("counts", "rate"), ...)

\S4method{plot}{PeakModel,missing}(x, y, ...)

\S4method{plot}{CalibrationCurve,missing}(x, y, ...)

\S4method{plot}{CalibrationCurve,DoseRate}(x, y, ...)
}
\arguments{
\item{x, y}{Objects to be plotted.}

\item{xaxis, yaxis}{A \code{\link{character}} string specifying the data to
be plotted along each axis.}

\item{...}{Currently not used.}

\item{select}{A \code{\link{numeric}} or \code{\link{character}} vector
giving the selection of the spectrum that are drawn.}

\item{facet}{A \code{\link{logical}} scalar: should a matrix of panels
defined by spectrum be drawn?}
}
\description{
Plot
}
\details{
TODO
}
\examples{
# Import CNF files
dir <- system.file("extdata/cerege/", package = "gamma")
spectra <- read(dir)

# Plot the spectrum named 'BRIQUE'
plot(spectra, select = "BRIQUE", facet = TRUE) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts") +
  ggplot2::theme_bw()

# Plot the first three spectra
plot(spectra, select = 1:3, facet = TRUE) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts") +
  ggplot2::theme_bw()
}
\author{
N. Frerebeau
}
