% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/predict.R
\docType{methods}
\name{predict}
\alias{predict}
\alias{predict-method}
\alias{predict,CalibrationCurve-method}
\title{Gamma dose rate}
\usage{
\S4method{predict}{CalibrationCurve}(object, spectra, epsilon = 0.03,
  ...)
}
\arguments{
\item{object}{An object of class \linkS4class{CalibrationCurve}.}

\item{spectra}{An optional object of class \linkS4class{GammaSpectra} in
which to look for variables with which to predict. If omitted, the fitted
values are used.}

\item{epsilon}{A \code{\link{numeric}} value.}

\item{...}{Currently not used.}
}
\description{
Predict in-situ gamma dose rate.
}
\examples{
# Import CNF files for calibration
spc_dir <- system.file("extdata/crp2a/calibration", package = "gamma")
spc <- read(spc_dir, skip = TRUE)
## Select 'BRIQUE', 'C341', 'C347', 'GOU', 'LMP', 'MAZ' and 'PEP' to build the curve
spc_calib <- spc[c("BRIQUE", "C341", "C347", "GOU", "LMP", "MAZ", "PEP")]

# Set dose rate values and errors for each spectrum
known_doses <- list(
  reference = c("BRIQUE", "C341", "C347", "GOU", "LMP", "MAZ", "PEP"),
  dose_value = c(1986, 850, 1424, 1575, 642, 1141, 2538),
  dose_error = c(36, 21, 24, 17, 18, 12, 112)
)
## Coerce to a 'DoseRate' object
known_doses <- methods::as(known_doses, "DoseRate")

# Build the calibration curve
calib_curve <- fit(
  spc_calib,
  doses = known_doses,
  noise = c(value = 1190, error = 1),
  range = c(200, 2800)
)

# Check the linear model
summary(calib_curve[["model"]])

# Plot the curve
plot(calib_curve) +
  ggplot2::labs(x = "Signal", y = "Dose rate [µGy/y]")

# Estimate gamma dose rates
(dose_rate <- predict(calib_curve, spc_calib))

# Plot the estimated values
plot(calib_curve, dose_rate) +
  ggplot2::labs(x = "Signal", y = "Dose rate [µGy/y]")
}
\author{
N. Frerebeau
}
