% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/doserate.R
\docType{methods}
\name{predict_dose}
\alias{predict_dose}
\alias{predict_dose-method}
\alias{predict_dose,CalibrationCurve,missing-method}
\alias{predict_dose,CalibrationCurve,GammaSpectrum-method}
\alias{predict_dose,CalibrationCurve,GammaSpectra-method}
\title{Gamma dose rate}
\usage{
predict_dose(object, spectrum, ...)

\S4method{predict_dose}{CalibrationCurve,missing}(object, epsilon = 0,
  simplify = FALSE, ...)

\S4method{predict_dose}{CalibrationCurve,GammaSpectrum}(object, spectrum,
  epsilon = 0, simplify = FALSE, ...)

\S4method{predict_dose}{CalibrationCurve,GammaSpectra}(object, spectrum,
  epsilon = 0, simplify = FALSE, ...)
}
\arguments{
\item{object}{A \linkS4class{CalibrationCurve} object.}

\item{spectrum}{An optional \linkS4class{GammaSpectrum} or
\linkS4class{GammaSpectra} object in which to look for variables with which
to predict. If omitted, the fitted values are used.}

\item{...}{Currently not used.}

\item{epsilon}{A \code{\link{numeric}} value giving an extra error term
introduced by the energy calibration of the spectrum.}

\item{simplify}{A \code{\link{logical}} scalar: should the result be
simplified to a matrix? If \code{FALSE} (default), returns a list.}
}
\value{
If \code{simplify} is \code{FALSE} returns a list of length-two numeric
 vectors (default), else returns a matrix.
}
\description{
Predict in-situ gamma dose rate.
}
\examples{
# Import CNF files for calibration
spc_dir <- system.file("extdata/BDX100/calibration", package = "gamma")
spectra <- read(spc_dir)

# Set dose rate values and errors for each spectrum
data("clermont")
set_dose(spectra) <- clermont[, c("gamma", "gamma_error")]

# Build the calibration curve
calib_curve <- fit_dose(
  spectra,
  noise = c(25279.63171, 1.66235),
  range = c(165, 2800)
)

# Check the linear model
summary(calib_curve[["model"]])

# Plot the curve
plot(calib_curve) +
  ggplot2::labs(x = "Signal", y = "Dose rate [µGy/y]")

# Estimate gamma dose rates
(dose_rate <- predict_dose(calib_curve, spectra))
}
\seealso{
Other dose rate: \code{\link{fit_dose}}
}
\author{
N. Frerebeau
}
\concept{dose rate}
