% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/baseline.R
\docType{methods}
\name{processBaseline}
\alias{processBaseline}
\alias{estimateBaseline}
\alias{estimateBaseline-method}
\alias{removeBaseline}
\alias{removeBaseline-method}
\alias{estimateBaseline,GammaSpectra-method}
\alias{estimateBaseline,GammaSpectrum-method}
\alias{removeBaseline,GammaSpectra-method}
\alias{removeBaseline,GammaSpectrum-method}
\title{Baseline estimation and removal}
\usage{
estimateBaseline(object, ...)

removeBaseline(object, ...)

\S4method{estimateBaseline}{GammaSpectra}(object, method = c("SNIP"),
  LLS = FALSE, ...)

\S4method{estimateBaseline}{GammaSpectrum}(object, method = c("SNIP"),
  LLS = FALSE, ...)

\S4method{removeBaseline}{GammaSpectra}(object, ...)

\S4method{removeBaseline}{GammaSpectrum}(object, ...)
}
\arguments{
\item{object}{A \code{\link{numeric}} vector.}

\item{...}{Extra parameters passed to other methods.}

\item{method}{TODO.}

\item{LLS}{A \code{\link{logical}} scalar: should the LLS operator be applied
on \code{x} before employing SNIP algorithm?}
}
\value{
\code{estimateBaseline} returns a four columns data frame or a list
 of data frame.

 \code{removeBaseline} returns a modified \linkS4class{GammaSpectrum} or
 \linkS4class{GammaSpectra} object (same as \code{object}, with background
 removed).
}
\description{
Baseline estimation and removal
}
\examples{
# Import a CNF file
spc_dir <- system.file("extdata/test1.cnf", package = "gamma")
spc_dir <- system.file("extdata/cerege", package = "gamma")
spc <- read(spc_dir)[[5]]

# Estimate baseline
baseline <- estimateBaseline(spc)

# Plot spectrum + baseline
plot(spc, baseline) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts")

# Remove baseline
spc_clean <- removeBaseline(spc)

# Detect peaks
peak_index <- findPeaks(spc_clean, span = NULL)

plot(spc_clean, select = 1:3, facet = TRUE) +
  ggplot2::labs(x = "Energy [keV]", y = "Counts") +
  ggplot2::geom_vline(xintercept = peak_index$energy, colour = "red") +
  ggplot2::theme_bw()
}
\references{
Morháč, M., Kliman, J., Matoušek, V., Veselský, M. and Turzo, I. (1997).
 Background elimination methods for multidimensional gamma-ray spectra.
 \emph{Nuclear Instruments and Methods in Physics Research Section A:
 Accelerators, Spectrometers, Detectors and Associated Equipment}, 401(1), p. 113-132.
 DOI: \href{https://doi.org/10.1016/S0168-9002(97)01023-1}{10.1016/S0168-9002(97)01023-1}

 Morháč, M. and Matoušek, V. (2008). Peak Clipping Algorithms for Background
 Estimation in Spectroscopic Data. \emph{Applied Spectroscopy}, 62(1), p. 91-106.
 DOI: \href{https://doi.org/10.1366/000370208783412762}{10.1366/000370208783412762}

 Ryan, C. G., Clayton, E., Griffin, W. L., Sie, S. H. and Cousens, D. R.
 (1988). SNIP, a statistics-sensitive background treatment for the
 quantitative analysis of PIXE spectra in geoscience applications.
 \emph{Nuclear Instruments and Methods in Physics Research Section B:
 Beam Interactions with Materials and Atoms}, 34(3), p. 396-402.
 DOI: \href{https://doi.org/10.1016/0168-583X(88)90063-8}{10.1016/0168-583X(88)90063-8}
}
\author{
N. Frerebeau
}
