% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/read.R
\docType{methods}
\name{read}
\alias{read}
\alias{read-method}
\alias{read,character-method}
\title{Data input}
\usage{
read(file, ...)

\S4method{read}{character}(file, ...)
}
\arguments{
\item{file}{A \code{\link{character}} string giving the path of files to be
imported.}

\item{...}{Extra parameters passed to \code{\link[rxylib]{read_xyData}}.}
}
\value{
An object of class \linkS4class{GammaSpectra}.
}
\description{
Reads a gamma dose rate file.
}
\details{
TODO
}
\note{
\emph{Only works with Canberra CNF.}
}
\examples{
# Import a Canberra CNF file
file <- system.file("extdata/gamma.cnf", package = "gamma")
(spectrum <- read(file))

# Import all CNF files in a given directory
dir <- system.file("extdata/cerege/", package = "gamma")
(spectra <- read(dir))
}
\seealso{
\link[rxylib]{read_xyData}
}
\author{
N. Frerebeau
}
