% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/read.R
\docType{methods}
\name{read}
\alias{read}
\alias{read-method}
\alias{read,character-method}
\title{Data input}
\usage{
read(file, ...)

\S4method{read}{character}(file, extensions = c("cnf", "tka"),
  skip = NULL, ...)
}
\arguments{
\item{file}{A \code{\link{character}} string giving the path of files to be
imported.}

\item{...}{Extra parameters passed to \code{\link[rxylib]{read_xyData}}.}

\item{extensions}{A \code{\link{character}} vector specifying the possible
file extensions. It must be one or more of "\code{cnf}", "\code{tka}".}

\item{skip}{A list of \code{\link{numeric}} vectors or of
\code{\link{logical}} scalars (re-cycled to the number of files to be
imported).
If \code{numeric}, the corresponding chanels of the data file will be
skipped. If \code{TRUE}, it will try to automatically process the
spectrum: all channels before the first maximum will be skipped. If
\code{NULL} (the default), all data are imported.}
}
\value{
An object of class \linkS4class{GammaSpectra}.
}
\description{
Reads a gamma ray spectrum file.
}
\note{
\emph{Only supports Canberra CNF and TKA files.}
}
\examples{
# Import a Canberra CNF file
cnf_file <- system.file("extdata/test.cnf", package = "gamma")
(cnf_spectrum <- read(cnf_file))

# Import a TKA file
tka_file <- system.file("extdata/test.tka", package = "gamma")
(tka_spectrum <- read(tka_file))

# Import all files in a given directory
# Skip the 35 first chanels
spc_dir <- system.file("extdata/crp2a/calibration", package = "gamma")
(spc_spectra <- read(spc_dir, skip = 1:35))
}
\seealso{
\link[rxylib]{read_xyData}
}
\author{
N. Frerebeau
}
