% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/read.R
\docType{methods}
\name{read}
\alias{read}
\alias{read-method}
\alias{read,character-method}
\title{Data input}
\usage{
read(file, ...)

\S4method{read}{character}(file, extensions = c("cnf", "tka"),
  skip = NULL, ...)
}
\arguments{
\item{file}{A \code{\link{character}} string giving the path of files to be
imported.}

\item{...}{Extra parameters passed to \code{\link[rxylib]{read_xyData}}.}

\item{skip}{A \code{\link{numeric}} vector.}
}
\value{
An object of class \linkS4class{GammaSpectra}.
}
\description{
Reads a gamma ray spectrum file.
}
\details{
\emph{Only works with Canberra CNF.}
}
\examples{
# Import a Canberra CNF file
cnf_file <- system.file("extdata/test.cnf", package = "gamma")
(cnf_spectrum <- read(cnf_file))

# Import a TKA file
tka_file <- system.file("extdata/test.tka", package = "gamma")
(tka_spectrum <- read(tka_file))

# Import all files in a given directory
# Skip the 35 first chanels
spc_dir <- system.file("extdata/crp2a/", package = "gamma")
(spc_spectra <- read(spc_dir, skip = 1:35))
}
\seealso{
\link[rxylib]{read_xyData}
}
\author{
N. Frerebeau
}
