% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/read.R
\docType{methods}
\name{read}
\alias{read}
\alias{read-method}
\alias{read,character-method}
\title{Data input}
\usage{
read(file, ...)

\S4method{read}{character}(file, extensions = c("cnf", "tka"), ...)
}
\arguments{
\item{file}{A \code{\link{character}} string giving the path of files to be
imported.}

\item{...}{Extra parameters to be passed to
\code{\link[rxylib]{read_xyData}}.}

\item{extensions}{A \code{\link{character}} vector specifying the possible
file extensions. It must be one or more of "\code{cnf}", "\code{tka}".}
}
\value{
A \linkS4class{GammaSpectra} object if more than one spectrum are imported
 at once, else a \linkS4class{GammaSpectrum} object.
}
\description{
Reads a gamma ray spectrum file.
}
\note{
\emph{Only supports Canberra CNF and TKA files.}
}
\examples{
# Import a Canberra CNF file
cnf_file <- system.file("extdata/test_CNF.cnf", package = "gamma")
(cnf_spectrum <- read(cnf_file))

# Import a TKA file
tka_file <- system.file("extdata/test_TKA.tka", package = "gamma")
(tka_spectrum <- read(tka_file))

# Import all files in a given directory
# Skip the 35 first chanels
spc_dir <- system.file("extdata/", package = "gamma")
(spc_spectra <- read(spc_dir))
}
\seealso{
\link[rxylib]{read_xyData}
}
\author{
N. Frerebeau
}
