% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/simulate.R
\docType{methods}
\name{simulate_spectrum}
\alias{simulate_spectrum}
\alias{simulate_spectrum-method}
\alias{simulate_spectrum,numeric,numeric,numeric-method}
\title{Simulate a Gamma-Ray Spectrum}
\usage{
simulate_spectrum(K, U, Th, ...)

\S4method{simulate_spectrum}{numeric,numeric,numeric}(K, U, Th,
  energy = c(1, 3200), n = 1024, ...)
}
\arguments{
\item{K, Th, U}{A length-one \code{\link{numeric}} vector giving the K-U-Th
ratio that caracterize a natural spectrum.}

\item{...}{Currently not used.}

\item{energy}{A length-two \code{\link{numeric}} vector giving the energy
range (in keV).}

\item{n}{An \code{\link{integer}} giving the number of chanel.}
}
\value{
A \linkS4class{GammaSpectrum} object.
}
\description{
Rough simulation of a gamma-ray spectrum.
}
\author{
N. Frerebeau
}
\keyword{internal}
