% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/smooth.R
\docType{methods}
\name{smooth}
\alias{smooth}
\alias{smooth-method}
\alias{smooth,GammaSpectra-method}
\alias{smooth,GammaSpectrum-method}
\title{Smooth}
\usage{
smooth(object, ...)

\S4method{smooth}{GammaSpectra}(object, method = c("unweighted",
  "weighted", "savitzky"), m = 3, p = 2)

\S4method{smooth}{GammaSpectrum}(object, method = c("unweighted",
  "weighted", "savitzky"), m = 3, p = 2)
}
\arguments{
\item{object}{A \linkS4class{GammaSpectrum} or \linkS4class{GammaSpectra}
object.}

\item{...}{Currently not used.}

\item{method}{A \code{\link{character}} string specifying the smoothing
method to be used. It must be one of "\code{unweighted}" (default),
"\code{weighted}" or "\code{savitzky}" (see details).
Any unambiguous substring can be given.}

\item{m}{An odd \code{\link{integer}} scalar giving the number of adjacent
points to be used.}

\item{p}{An \code{\link{integer}} scalar giving the polynomial degree.
Only used if \code{method} is \code{savitzky}.}
}
\value{
A \linkS4class{GammaSpectrum} or \linkS4class{GammaSpectra} object.
}
\description{
Smoothes a spectrum.
}
\details{
The following smoothing methods are available:
 \describe{
  \item{unweighted}{Unweighted sliding-average or rectangular smooth.
  It replaces each point in the signal with the average of \code{m} adjacent
  points.}
  \item{weighted}{Weighted sliding-average or triangular smooth.
  It replaces each point in the signal with the weighted mean of \code{m}
  adjacent points.}
  \item{savitzky}{Savitzky-Golay filter.}
 }
 There will be \eqn{(m - 1) / 2} points both at the beginning and at the end
 of the spectrum for which a complete \code{m}-width smooth cannot be
 calculated. To prevent data loss, progressively smaller smooths are used at
 the ends of the spectrum if \code{method} is \code{unweighted} or
 \code{weighted}. If the Savitzky-Golay filter is used, the original
 \eqn{(m - 1) / 2} points at the ends of the spectrum are preserved.
}
\examples{
library(magrittr)

# Import CNF files
spc_file <- system.file("extdata/test_CNF.cnf", package = "gamma")
spectrum <- read(spc_file, skip = TRUE)

# Plot spectrum
spectrum \%>\%
  removeBaseline() \%>\%
  plot()

# Rectangular smooth
spectrum \%>\%
  smooth(method = "unweighted", m = 3) \%>\%
  removeBaseline() \%>\%
  plot()

# Triangular smooth
spectrum \%>\%
  smooth(method = "weighted", m = 5) \%>\%
  removeBaseline() \%>\%
  plot()

# Savitzky–Golay
spectrum \%>\%
  smooth(method = "savitzky", m = 5, p = 2) \%>\%
  removeBaseline() \%>\%
  plot()
}
\references{
Gorry, P. A. (1990). General Least-Squares Smoothing and Differentiation by
 the Convolution (Savitzky-Golay) Method. \emph{Analytical Chemistry}, 62(6),
 p. 570-573.
 DOI: \href{https://doi.org/10.1021/ac00205a007}{10.1021/ac00205a007}.

 Savitzky, A. and Golay, M. J. E. (1964). Smoothing and Differentiation of
 Data by Simplified Least Squares Procedures. \emph{Analytical Chemistry},
 36(8), p. 1627-1639.
 DOI: \href{https://doi.org/10.1021/ac60214a047}{10.1021/ac60214a047}.
}
\seealso{
Other signal processing: \code{\link{integrateSignal}},
  \code{\link{peaks}}, \code{\link{processBaseline}}
}
\author{
N. Frerebeau
}
\concept{signal processing}
