% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/signal.R
\docType{methods}
\name{smooth_signal}
\alias{smooth_signal}
\alias{smooth_signal-method}
\alias{smooth_signal,GammaSpectrum-method}
\alias{smooth_signal,GammaSpectra-method}
\title{Smooth}
\usage{
smooth_signal(object, ...)

\S4method{smooth_signal}{GammaSpectrum}(
  object,
  method = c("unweighted", "weighted", "savitzky"),
  m = 3,
  p = 2
)

\S4method{smooth_signal}{GammaSpectra}(
  object,
  method = c("unweighted", "weighted", "savitzky"),
  m = 3,
  p = 2
)
}
\arguments{
\item{object}{A \linkS4class{GammaSpectrum} or \linkS4class{GammaSpectra}
object.}

\item{...}{Currently not used.}

\item{method}{A \code{\link{character}} string specifying the smoothing
method to be used. It must be one of "\code{unweighted}" (default),
"\code{weighted}" or "\code{savitzky}" (see details).
Any unambiguous substring can be given.}

\item{m}{An odd \code{\link{integer}} giving the number of adjacent
points to be used.}

\item{p}{An \code{\link{integer}} giving the polynomial degree.
Only used if \code{method} is "\code{savitzky}".}
}
\value{
A \linkS4class{GammaSpectrum} or \linkS4class{GammaSpectra} object.
}
\description{
Smoothes intensities.
}
\details{
The following smoothing methods are available:
 \describe{
  \item{unweighted}{Unweighted sliding-average or rectangular smooth.
  It replaces each point in the signal with the average of \eqn{m} adjacent
  points.}
  \item{weighted}{Weighted sliding-average or triangular smooth.
  It replaces each point in the signal with the weighted mean of \eqn{m}
  adjacent points.}
  \item{savitzky}{Savitzky-Golay filter. This method is based on the
  least-squares fitting of polynomials to segments of \eqn{m} adjacent
  points.}
 }
 There will be \eqn{(m - 1) / 2} points both at the beginning and at the end
 of the spectrum for which a complete \eqn{m}-width smooth cannot be
 calculated. To prevent data loss, progressively smaller smooths are used at
 the ends of the spectrum if \code{method} is \code{unweighted} or
 \code{weighted}. If the Savitzky-Golay filter is used, the original
 \eqn{(m - 1) / 2} points at the ends of the spectrum are preserved.
}
\examples{
# Import CNF files
spc_file <- system.file("extdata/test_LaBr.CNF", package = "gamma")
spectrum <- read(spc_file)
spectrum <- slice_signal(spectrum)

# Plot spectrum
spc_clean <- remove_baseline(spectrum)
plot(spc_clean)

# Rectangular smooth
spc_unweighted <- smooth_signal(spectrum, method = "unweighted", m = 3)
spc_unweighted_clean <- remove_baseline(spc_unweighted)
plot(spc_unweighted_clean)

# Triangular smooth
spc_weighted <- smooth_signal(spectrum, method = "weighted", m = 5)
spc_weighted_clean <- remove_baseline(spc_weighted)
plot(spc_weighted_clean)

# Savitzky–Golay
spc_savitzky <- smooth_signal(spectrum, method = "savitzky", m = 5, p = 2)
spc_savitzky_clean <- remove_baseline(spc_savitzky)
plot(spc_savitzky_clean)
}
\references{
Gorry, P. A. (1990). General Least-Squares Smoothing and Differentiation by
 the Convolution (Savitzky-Golay) Method. \emph{Analytical Chemistry}, 62(6),
 p. 570-573.
 DOI: \href{https://doi.org/10.1021/ac00205a007}{10.1021/ac00205a007}.

 Savitzky, A. & Golay, M. J. E. (1964). Smoothing and Differentiation of
 Data by Simplified Least Squares Procedures. \emph{Analytical Chemistry},
 36(8), p. 1627-1639.
 DOI: \href{https://doi.org/10.1021/ac60214a047}{10.1021/ac60214a047}.
}
\seealso{
Other signal processing: 
\code{\link{baseline}},
\code{\link{find_peaks}()},
\code{\link{integrate_signal}()},
\code{\link{slice_signal}()},
\code{\link{stabilize_signal}()}
}
\author{
N. Frerebeau
}
\concept{signal processing}
