% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/stabilize-methods.R
\docType{methods}
\name{stabilize}
\alias{stabilize}
\alias{stabilize-method}
\alias{stabilize,GammaSpectrum-method}
\title{Transform Intensities}
\usage{
stabilize(object, ...)

\S4method{stabilize}{GammaSpectrum}(object, transformation, ...)
}
\arguments{
\item{object}{A \linkS4class{GammaSpectrum} object.}

\item{...}{Extra arguments to be passed to \code{transformation}.}

\item{transformation}{A \code{\link{function}} that takes a numeric vector as
argument and returns a numeric vector.}
}
\value{
A new \linkS4class{GammaSpectrum} object with transformed
 intensities.
}
\description{
Transform Intensities
}
\seealso{
Other signal processing: \code{\link{integrateSignal}},
  \code{\link{peaks}}, \code{\link{processBaseline}},
  \code{\link{smooth}}
}
\author{
N. Frerebeau
}
\concept{signal processing}
