% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/show.R
\docType{methods}
\name{summarise}
\alias{summarise}
\alias{summarise-method}
\alias{summarise,GammaSpectrum-method}
\alias{summarise,GammaSpectra-method}
\title{Summarize}
\usage{
summarise(object, ...)

\S4method{summarise}{GammaSpectrum}(object)

\S4method{summarise}{GammaSpectra}(object)
}
\arguments{
\item{object}{A \linkS4class{GammaSpectrum} or  \linkS4class{GammaSpectra}
object.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{data.frame}}
}
\description{
Summarize
}
\author{
N. Frerebeau
}
\concept{summarise}
